% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{adjust_species_attributes}
\alias{adjust_species_attributes}
\title{Harmonize species attributes with TerrADat needs}
\usage{
adjust_species_attributes(data, fail_on_missing = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{Data frame. Long format LPI data with any of the required variables: \code{"GrowthHabit"}, \code{"GrowthHabitSub"}, \code{"Duration"}, \code{"Family"}, \code{"HigherTaxon"}, \code{"Nonnative"}, \code{"Invasive"}, \code{"Noxious"}, \code{"SpecialStatus"}, \code{"Photosynthesis"}, \code{"PJ"}, \code{"chckbox"}.}

\item{fail_on_missing}{Logical. If \code{FALSE} then any modifications which depend on expected variables which are \emph{not} present will be skipped. If \code{TRUE}, missing expected variables will cause an error. Defaults to \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
The original data frame with modified and additional variables for use in \code{lpi_calc()}.
}
\description{
This function is intended to be applied to LPI data which have been attributed using \code{\link[terradactyl:species_join]{species_join()}} with BLM Terrestrial AIM tblNationalPlants and tblStateSpecies read in via \code{\link[terradactyl:species_read_aim]{species_read_aim()}}.
It modifies the values of some variables to clean values (e.g., converting all capitalization and hyphenation schemes for the string \code{"non-woody"} to \code{"NonWoody"}) to include relevant records in calculations and to prevent the need to rename indicator outputs.
It also adds variables for use in derived indicators, e.g., creating the variable Plant which contains the value \code{"Plant"} only for records associated with vascular plants based on the variables GrowthHabit and code).
Modifications are done only if the necessary variables already exist, e.g., modifications or additions which use the variable Duration will only be attempted if Duration exists in the provided data.
}
\details{
Special definitions for classifications (e.g., how records are identified as corresponding to a lichen) are from \code{\link[terradactyl:lpi_indicator_definitions]{lpi_indicator_definitions()}}. This is the source of definitions referenced without explicit values in the details below.
The following changes are made (if possible):

\itemize{
  \item Duration
    \itemize{
   \item Any value matching the case-insensitive regex \code{"perennial"} becomes \code{"Peren"}.
   \item Any value matching the case-insensitive regex \code{"annual|biennial"} becomes \code{"Ann"}.
   \item \code{NA}s become \code{"duration_irrelevant"}.
   \item Any other values are left unaltered.
    }

  \item GrowthHabit
    \itemize{
   \item Any value matching the case-insensitive regex \code{"^non-?woody$"} becomes \code{"NonWoody"}.
   \item Any value matching the case-insensitive regex \code{"^non-?vascular$"} becomes \code{"Nonvascular"}.
   \item \code{NA}s become \code{"growthhabit_irrelevant"}.
   \item Any other values are left unaltered.
   }

  \item GrowthHabitSub
    \itemize{
   \item Any value matching the case-insensitive regex \code{"forb"} becomes \code{"Forb"}.
   \item Any value matching the case-insensitive regex \code{"^sub-?shrub$"} becomes \code{"SubShrub"}.
   \item Any value where the associated GrowthHabit value matches one of the case-insensitive regexes \code{"^non-?vascular$"}, \code{"^moss$"}, or \code{"^lichen$"} becomes \code{"growthhabitsub_irrelevant"}.
   \item \code{NA}s become \code{"growthhabitsub_irrelevant"}.
   \item All other values are left unaltered.
   }

  \item Plant
    \itemize{
   \item This is a new variable not expected in the input species attributes and is used to calculate basal cover by vascular plants, total foliar cover by vascular plants, and making sure plant properties aren't assigned to non-plant records in other variables that are made/modified by this function.
   \item The value of this variable will be "Plant" only if ALL of the following criteria are true, otherwise this value with be \code{NA}:
     \itemize{
     \item The associated value in GrowthHabitSub is \emph{not} \code{"growthhabitsub_irrelevant"}.
     \item The associated value in GrowthHabit is \emph{not} \code{"Nonvascular"}.
     \item The associated value in code consists of three or more characters. As of January 2026, this is still assumed to be a difference between valid species codes and all other types of values found in the code variable.
     }
    }

  \item ShrubSucculent
    \itemize{
   \item This is a new variable added to calculate indicators specific to shrubs, subshrubs, and succulents.
   \item The value of this variable will be \code{"ShrubSucculent"} if the associated value in GrowthHabitSub matches the case-insensitive regex \code{"shrub|succulent"}.
   \item Otherwise, this value will be \code{NA}.
   }

  \item Litter
    \itemize{
   \item This is a new variable added to calculate indicators specific to litter types.
   \item Any record where value of the code variable is in the set of herbaceous litter codes is populated with \code{"HerbLitter"}.
   \item Any record where value of the code variable is in the set of woody litter codes is populated with \code{"WoodyLitter"}.
   \item Any record where value of the code variable is in the set of embedded litter codes is populated with \code{"EmbLitter"}.
   \item Any record where code is any other value is populated with \code{"litter_irrelevant"}.
   }

  \item TotalLitter
    \itemize{
   \item This is a new variable added to calculate total litter cover.
   \item Any record where value of the code variable in the set of herbaceous, woody, or embedded litter codes is populated with \code{"TotalLitter"}.
   \item Any record where code is any other value is populated with \code{NA}.
   }

  \item Biocrust
    \itemize{
   \item This is a new variable added to calculate biocrust cover.
   \item Any record where value of the code variable in the set of biocrust codes is populated with \code{"Biocrust"}.
   \item Any record where code is any other value is populated with \code{NA}.
   }

  \item Lichen
    \itemize{
   \item This is a new variable added to calculate indicators specific to lichen and lichen-adjacent records.
   \item Any record where value of the code variable is in the set of lichen codes is populated with \code{"Lichen"}.
   \item Any record where value of the code variable is in the set of vagrant lichen codes is populated with \code{"VagrLichen"}.
   \item Any record where value of the code variable is in the set of cyanobacteria codes is populated with \code{"Cyanobacteria"}.
   \item Any record where code is any other value is populated with \code{"litter_irrelevant"}.
   }

  \item PJ
    \itemize{
   \item This is either an existing variable that is modified or a new variable added to calculate pinyon-juniper cover.
   \item If PJ is an existing variable
     \itemize{
       \item Any record where value of PJ is \code{TRUE} is replaced with \code{"PJ"}.
       \item All other records are replaced with \code{NA}.
     }

   \item If PJ is \emph{not} an existing variable
     \itemize{
       \item Any record where value of the code variable in the set of PJ codes is populated with \code{"PJ"}.
       \item Any record where code is any other value is populated with \code{NA}.
     }
     }

  \item Rock
    \itemize{
   \item This is a new variable added to enable the calculation of rock indicators.
   \item Any record where value of the code variable is in the set of rock fragment codes is populated with \code{"Rock"}.
   \item Otherwise this is populated with \code{NA}.
   }

  \item Duff
    \itemize{
   \item This is a new variable added to enable the calculation of duff indicators.
   \item Any record where value of the code variable is in the set of duff fragment codes is populated with \code{"Duff"}.
   \item Otherwise this is populated with \code{NA}.
   }

  \item Water
    \itemize{
   \item This is a new variable added to enable the calculation of water indicators.
   \item Any record where value of the code variable is in the set of water fragment codes is populated with \code{"Rock"}.
   \item Otherwise this is populated with \code{"water_irrelevant"}.
   }

  \item C3
    \itemize{
   \item This is a new variable added to enable the calculation of C3 photosynthetic pathway indicators because a single record may contain both \code{"C3"} and \code{"C4"} in the Photosynthesis variable.
   \item The value of this variable will be \code{"C3"} if the associated value in Photosynthesis matches the regex \code{"C3"}.
   \item Otherwise, the value will be \code{NA}.
   }

  \item C4
    \itemize{
   \item This is a new variable added to enable the calculation of C4 photosynthetic pathway indicators because a single record may contain both \code{"C3"} and \code{"C4"} in the Photosynthesis variable.
   \item The value of this variable will be \code{"C4"} if the associated value in Photosynthesis matches the regex \code{"C4"}.
   \item Otherwise, the value will be \code{NA}.
   }

  \item Grass
    \itemize{
   \item This is a new variable added to enable the calculation of grass-specific indicators.
   \item The value of this variable will be \code{"Grass"} if the associated value in Family is \code{"Poaceae"}.
   \item Otherwise, the value will be \code{NA}.
   }

  \item Conifer
    \itemize{
   \item This is a new variable added to enable the calculation of conifer-specific indicators.
   \item The value of this variable will be \code{"Conifer"} if the associated value in Family is one of the identified conifer families.
   \item Otherwise, the value will be \code{NA}.
   }

  \item ForbGraminoid
    \itemize{
   \item This is a new variable added to enable the calculation of forb-and-graminoid-specific indicators.
   \item The value of this variable will be \code{"ForbGraminoid"} if the associated value in GrowthHabitSub matches the regex \code{"(^((graminoid)|(grass))$)|forb"}.
   \item Otherwise, the value will be \code{NA}.
   }

  \item ForbGrass
    \itemize{
   \item This is a new variable added to enable the calculation of forb-and-grass-specific indicators.
   \item The value of this variable will be \code{"ForbGrass"} if the associated value in GrowthHabitSub matches the regex \code{"forb"} \emph{or} if the value in Family is \code{"Poaceae"}.
   \item Otherwise, the value will be \code{NA}.
   }

  \item SG_Group
    \itemize{
   \item This variable is expected to initially contain a string with multiple values corresponding to different localities and uses the value from SpeciesState to extract only the relevant substring with the regex produced with \code{paste0("(?<=((US)|(", SpeciesState, ")):)[A-z]+")}.
   \item If a substring was extracted and it contains the substring \code{"StaturePerennialGrass"}, that substring is replaced with \code{"PerenGrass"}, e.g., \code{"TallStaturePerennialGrass"} becomes \code{"TallPerenGrass"}.
   \item If there was no matching substring extracted and the corresponding GrowthHabitSub value is \code{"Shrub"}, the value is populated with \code{"NonSagebrushShrub"}.
   \item If there was no matching substring extracted and the corresponding GrowthHabitSub value is \emph{not} \code{"Shrub"}, the value is populated with \code{"Irrelevant"}.
   \item Otherwise, the value is populated with the extracted substring.
   }

  \item Moss
    \itemize{
   \item This is a new variable added to enable the calculation of moss-specific indicators.
   \item The value of this variable will be \code{"Moss"} if the associated value in code matches the regex \code{"^(M(OS{1,2})?\\d+)|(2?MOSS)$"} \emph{or} the value in code is \code{"M"} \emph{or} if the value in HigherTaxon is \code{"Moss"}.
   \item Otherwise, the value will be \code{"moss_irrelevant"}.
   }

  \item between_plant
    \itemize{
   \item This is a new variable added to enable the calculation of between-vascular-plants indicators.
   \item Any record where value of the code variable is in the set of herbaceous litter codes is populated with \code{"HerbLitter"}.
   \item Any record where value of the code variable is in the set of woody litter codes is populated with \code{"WoodyLitter"}.
   \item Any record where value of the code variable is in the set of embedded litter codes is populated with \code{"EmbLitter"}.
   \item Any record where value of the code variable is in the set of deposited soil codes is populated with \code{"DepSoil"}.
   \item Any record where value of the code variable is in the set of duff codes is populated with \code{"Duff"}.
   \item Any record where value of the code variable is in the set of water codes is populated with \code{"Water"}.
   \item Any record where value of the code variable is in the set of rock fragment codes is populated with \code{"Rock"}.
   \item Any record where value of the code variable is in the set of bare soil codes is populated with \code{"BareSoil"}.
   \item Any record where value of the code variable is in the set of lichen codes is populated with \code{"Lichen"}.
   \item Any record where value of the code variable is in the set of vagrant lichen codes is populated with \code{"VagrLichen"}.
   \item Any record where value of the code variable is in the set of cyanobacteria codes is populated with \code{"Cyanobacteria"}.
   \item Any record where value of the code variable matches the regex \code{"^(M(OS{1,2})?\\d+)|(2?MOSS)$"} \emph{or} the value in code is \code{"M"} \emph{or} if the value in HigherTaxon is \code{"Moss"} is populated with \code{"Moss"}.
   \item Any record where code is any other value is populated with \code{"between_plant_irrelevant"}.
   }

  \item Native
    \itemize{
   \item This is a new variable created for native-specific cover indicators.
   \item Any record where the value of Plant is \code{TRUE} and Nonnative is one of \code{"Native"}, \code{"NATIVE"}, \code{"native"}, or \code{NA} is populated with \code{"Native"}.
   \item Any record where the value of Plant is \code{TRUE} and Nonnative is \emph{not} one of \code{"Native"}, \code{"NATIVE"}, \code{"native"}, or \code{NA} is populated with \code{"NonNative"}.
   \item Otherwise, this is populated with \code{NA}.
   }

  \item Invasive
    \itemize{
   \item All values are converted to title case, e.g. \code{"invasive"} becomes \code{"Invasive"}.
   \item Any \code{NA} values are converted to \code{"NonInv"}.
   }

  \item Noxious
    \itemize{
   \item This variable is expected to initially contain a string with multiple values corresponding to different localities and uses the value from SpeciesState to identify only the relevant substring with the regex produced with \code{"(^|\\|)((", SpeciesState, ")|(US))"}.
   \item If the regex was detected, the value is populated with \code{"Noxious"}.
   \item Otherwise, the value is populated with \code{"noxious_irrelevant"}.
   }
   }
}
