% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AERO.R
\name{aero_gap}
\alias{aero_gap}
\title{Calculate AERO Inputs}
\usage{
aero_gap(gap_tall, height_tall)
}
\arguments{
\item{gap_tall}{Table. Gap data in tall format}

\item{height_tall}{Table. Height data in tall format}
}
\description{
Calculate AERO Inputs
}
