% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{build.PK}
\alias{build.PK}
\title{Create valid PrimaryKey values}
\usage{
build.PK(plots)
}
\arguments{
\item{plots}{Data frame. The contents of \code{tblPlots} in TerrADat or an equivalent with variables named "SURVEY", "STATE", "COUNTY", "PSU", and "POINT".}
}
\value{
The data frame \code{plots} with a new variable "PrimaryKey" added containing the PrimaryKey value for each plot in the data frame.
}
\description{
The PrimaryKey values in TerrADat are strings created by pasting/concatenating several other values in the \code{tblPlots} table.
}
