% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R
\name{build_terradat_indicators}
\alias{build_terradat_indicators}
\title{Calculate the full set of standard Terrestrial AIM Database indicators}
\usage{
build_terradat_indicators(
  header,
  dsn,
  species_file,
  species_code_var = "SpeciesCode",
  lpi_tall = NULL,
  gap_tall = NULL,
  height_tall = NULL,
  spp_inventory_tall = NULL,
  soil_stability_tall = NULL,
  digits = 6,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{header}{Data frame or character string. The data to be provided as the argument \code{header} to any indicator calculation functions that require it. If this is a character string, it must point to the .Rdata file containing the data.}

\item{dsn}{Character string. The filepath to the geodatabase containing data. Passed to indicator calculation functions that require the argument \code{dsn}.}

\item{species_file}{Data frame or character string. The data to be provided as the argument \code{species_file} to any indicator calculation functions that require it. If this is a character string, it must point to the CSV or GDB file containing the data. This should almost always be to a geodatabase containing tblNationalPlants and tblStateSpecies.}

\item{species_code_var}{Character string. The name of the variable in the species characteristics that contain the species codes. Defaults to \code{"SpeciesCode"}.}

\item{lpi_tall}{Data frame or character string. The data to be provided as the argument \code{lpi_tall} to any indicator calculation functions that require it. If this is a character string, it must point to the .Rdata file containing the data.}

\item{gap_tall}{Data frame or character string. The data to be provided as the argument \code{gap_tall} to any indicator calculation functions that require it. If this is a character string, it must point to the .Rdata file containing the data.}

\item{height_tall}{Data frame or character string. The data to be provided as the argument \code{height_tall} to any indicator calculation functions that require it. If this is a character string, it must point to the .Rdata file containing the data.}

\item{spp_inventory_tall}{Data frame or character string. The data to be provided as the argument \code{spp_inventory_tall} to any indicator calculation functions that require it. If this is a character string, it must point to the .Rdata file containing the data.}

\item{soil_stability_tall}{Data frame or character string. The data to be provided as the argument \code{soil_stability_tall} to any indicator calculation functions that require it. If this is a character string, it must point to the .Rdata file containing the data.}

\item{...}{Optional are filtering statements. These will be passed to \code{dplyr::filter()} to applied to \code{header} to restrict the calculations.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A data frame with all standard TerrADat indicators in a format matching TerrADat.
}
\description{
A wrapper function for the *_calc() family of functions that produce the default TerrADat indicators.
}
