% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_qc.R
\name{check_terradat_data}
\alias{check_terradat_data}
\title{Check for uniqueness and orphaned records in raw AIM data.}
\usage{
check_terradat_data(
  dsn,
  data_types = c("lpi", "gap", "soilstability", "heights"),
  output_path = NULL,
  make_directory = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Character string. The filepath to the geodatabase containing the data to test. The data must be stored in the standard format, including table names, e.g., tblGapDetail.}

\item{data_types}{Character vector. The data type or types to test. Valid strings are \code{"lpi"}, \code{"gap"}, \code{"soilstability"}, and \code{"heights"}. Defaults to \code{c("lpi", "gap", "soilstability", "heights")}.}

\item{output_path}{Optional character string. The filepath to write all nonunique and orphaned records to. There will be one "nonuniques" CSV for each table where nonunique records were found and one "orphaned" CSV for each table where orphaned records were found. Filenames are formed from the name of the source table and the type of records, e.g. "tblGapHeader_nonuniques.CSV", "tblLLPIDetail_orphaned.CSV". No data will be written to CSV if this is \code{NULL}. Defaults to \code{NULL}.}

\item{make_directory}{Logical. If \code{TRUE} then the directory \code{output_path} will be created if it does not already exist. Defaults to \code{FALSE}.}
}
\value{
A named list of lists of data frame. The "nonuniques" list will contain named data frames with all records which were not unique. The "orphaned" list will contain named data frames with all records which didn't correspond to any records in their companion table. Data frames for tables which held only unique records or had no orphaned records will contain no values.
}
\description{
For a given geodatabase containing AIM (or AIM-compatible) data, check the request data types for uniqueness and orphaned records. This is a wrapper for check_uniqueness() and check_orphaned_records() which you can use if you have nonstandard data.
}
