% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_qc.R
\name{check_uniqueness}
\alias{check_uniqueness}
\title{Identify non-unique records in a data frame.}
\usage{
check_uniqueness(data, uid_variables)
}
\arguments{
\item{data}{Data frame. The data which will be checked for non-unique records.}

\item{uid_variables}{Character vector. One or more variable names corresponding to variables in the data frame. This must include only and all of the variables which when taken together should have only only one record for each combination of values. In a wide data set you would likely only have one variable but in a tall data set you would need to specify multiple. For example, in the case of tblGapDetail, this should be \code{c("PrimaryKey", "RecKey", "RecType", "Gap", "GapStart", "GapEnd")}.}
}
\value{
The original input \code{data} restricted to records which were non-unique with the added variable "id_nonunique_group" indicating which records are duplicates of each other.
}
\description{
Given a data frame and a set of variables, return the records from the data frame which do not have a unique combination of values in those variables.
This is intended to be used to find duplicated or impossible records.
}
\examples{
# example_data <- data.frame(key = rep(x = c(1, 2), times = 3),
#                            value1 = rep(x = c("terrestrial", "aquatic"), times = 3),
#                            value2 = 1:6)
# check_uniqueness(data = example_data,
#                  uid_variables = c("key",
#                                    "value1"))

}
