% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.indicators.R
\name{cover.indicators}
\alias{cover.indicators}
\alias{pct.cover.between.plant}
\alias{pct.cover.all.ground}
\alias{pct.cover.total.foliar}
\alias{pct.cover.bare.soil}
\alias{pct.cover.litter}
\title{Percent Cover Indicators}
\usage{
pct.cover.between.plant(lpi.tall, tall = FALSE, by.year = FALSE,
  by.line = FALSE)

pct.cover.all.ground(lpi.tall, tall = FALSE, by.year = FALSE,
  by.line = FALSE)

pct.cover.total.foliar(lpi.tall, tall = FALSE, by.year = FALSE,
  by.line = FALSE)

pct.cover.bare.soil(lpi.tall, tall = FALSE, by.year = FALSE,
  by.line = FALSE)

pct.cover.litter(lpi.tall, tall = FALSE, by.year = FALSE, by.line = FALSE)
}
\arguments{
\item{lpi.tall}{A tall/long-format data frame. Use the data frame \code{"layers"} from the \code{gather.lpi()} output.}

\item{by.year}{Logical. If \code{TRUE} then results will be reported further grouped by year using the \code{DateModified} field from the data forms. Defaults to \code{FALSE}.}

\item{by.line}{Logical. If \code{TRUR} then results will be reported further grouped by line using the \code{LineID} and \code{LineKey} fields from the data forms. Defaults to \code{FALSE}.}
}
\value{
A \code{tbl} of either wide or tall format.
}
\description{
Calculate the percent cover  indicators by plot or line for variables or combinations of variables.This is a family of standard indicator variables to examine total foliar cover, bare soil, litter cover, and other ground cover indicators. To compute cover by species, growth habit and duration, or other custom line-point intercept combinations, see \code{pct.cover()}.
}
