% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{default_indicators_vars}
\alias{default_indicators_vars}
\title{List variable groupings used for default cover indicator calculation}
\usage{
default_indicators_vars(
  source,
  hit = c("any", "first", "basal"),
  verbose = FALSE
)
}
\arguments{
\item{source}{Character string. Must correspond to the desired output type. The only currently accepted value is \code{"terradat"}.}

\item{hit}{Optional character vector. Which hit types to include in the list. May contain one or more of \code{"any"}, \code{"first"}, \code{"basal"}. Defaults to \code{c("any", "first", "basal")}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A named list of one or more list of vectors of variable names.
}
\description{
This returns a list of variables to use as the \code{indicator_variables} argument in \code{\link[terradactyl:pct_cover]{pct_cover()}}. The names of the list correspond to values for the \code{hit} argument in \code{pct_cover()}.
}
