% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{default_lpi_indicators}
\alias{default_lpi_indicators}
\title{Default expected cover indicators by source}
\usage{
default_lpi_indicators(source)
}
\arguments{
\item{source}{Character string. Must correspond to the desired output type. The only currently accepted value is \code{"terradat"}.}
}
\value{
A vector of names corresponding to possible cover indicators calculated by \code{lpi_calc()}.
}
\description{
Get a vector of expected cover indicator names to be produced by \code{\link[terradactyl:lpi_calc]{lpi_calc()}} for a given data source.
}
