% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{demo_data}
\alias{demo_data}
\title{Load demo data for terradactyl examples}
\usage{
demo_data(type, verbose = FALSE)
}
\arguments{
\item{type}{Character string. This determines
which data set to load. Valid values are
\code{"lpi"}, \code{"gap"}, \code{"height"}, \code{"generic_species"},
\code{"lpi_species"}, \code{"rangeland_health"},
\code{"species_inventory"}, \code{"soil_stability"}, \code{"species_list"},
\code{"generic_species"}, and \code{"header"}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
Data frame.
}
\description{
The package terradactyl contains example data sets that can be
  used for testing or to check formats.

  The example data include indicators (the output from
  \code{build_indicators()}) and the long/tall-format data produced by the
  various gather functions, e.g. \code{gather_lpi()}.
}
\examples{
# Loading example Terrestrial AIM indicators produced by build_indicators()
demo_data(type = "indicators")

# Loading example long/tall-format line-point intercept data produced by
# gather_lpi()
demo_data(type = "lpi")

# Loading example gap data produced by gather_gap()
demo_data(type = "gap")
}
