% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dima.R
\name{fetch_api_single}
\alias{fetch_api_single}
\alias{fetch_api}
\title{Fetch DIMA data from the Landscape Data Commons API (requires internet connection)}
\usage{
fetch_api_single(api, endpoint, values = NULL, verbose = T)

fetch_api(api, endpoint, values = NULL)
}
\arguments{
\item{endpoint}{Character string or list. Name of the requested table e.g. "tblPlots". 
Provide a list of character strings to download more than one table.}

\item{values}{Optional. JSON query as character string. If a list of endpoints is provided, the filter column must be present in all requested tables.}

\item{verbose}{If true, print the URL of the requested record.}
}
\value{
A data frame containing DIMA data of the requested table, or a list
of data frames containing the requested tables.
}
\description{
Given a table name and optional field query, download DIMA data from the API.
}
\examples{
data_allplots <- fetch_api(endpoint = "tblPlots", values = NULL)

data_JERplots <- fetch_api(endpoint = "tblPlots", values = "ProjectKey=JER")

data_gap <- fetch_api(endpoint = list("tblGapHeader", "tblGapDetail"), 
                       values = "PrimaryKey=15050113465465692020-09-15")
}
