% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R
\name{gap_calc}
\alias{gap_calc}
\title{Calculate the standard Terrestrial AIM Database (TerrADat) Canopy Gap indicators}
\usage{
gap_calc(header = NULL, gap_tall, digits = 1, verbose = FALSE)
}
\arguments{
\item{header}{Optional data frame or character string. The metadata for the plots involved in the calculations, this is used to filter or subset the data being used for the calculations and must contain this must contain the variable PrimaryKey. If a character string, this must point to a CSV file containing the data. If \code{NULL} then no filtering will occur. Defaults to \code{NULL}.}

\item{gap_tall}{Data frame or character string. The long/tall-format gap data for the plots involved in the calculations. The format must match the output from \code{gather_gap()}. If a character string, this must point to a CSV file containing the data.}

\item{digits}{Integer. The number of decimal places that the output values will be rounded to. Values larger than \code{1} are not recommended because they will likely imply false precision. Defaults to \code{1}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A data frame matching the format of gap indicators in TerrADat.
}
\description{
This function calculates the full set of gap-derived indicators that are standard for TerrADat.

For any other gap indicators, use the underlying function \code{gap_cover()}.
}
