% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_all.R
\name{gather.all}
\alias{gather.all}
\title{Build tall tables}
\usage{
gather.all(dsn, out, method = c("gap", "lpi", "height",
  "species inventory", "soil stability", "rangeland health"),
  append = TRUE, source = c("AIM", "LMF"), file.type = "gdb")
}
\arguments{
\item{method}{String. Methods to append}

\item{append}{Logical. Append data to existing table}

\item{dsn.}{String. Filepath to data}
}
\description{
Build tall tables
}
