% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species}
\alias{species}
\alias{gather.species}
\title{Gather species data}
\usage{
gather.species(species.file, species.growth.habit.code = "GrowthHabitSub",
  growth.habit.file = "", growth.habit.code = "Code")
}
\arguments{
\item{species.file}{Character string. The full file path (including file extension) to the csv containing the species list. If NULL then the file from the provided geodatabase will be used.}

\item{species.growth.habit.code}{Character. The field name for the growth habit codes in the species file.}

\item{growth.habit.file}{Character string. The full file path (including file extension) to the csv containing the growth habit list. If NULL then the file from the provided geodatabase will be used.}

\item{growth.habit.code}{Character. The field name for the growth habit codes in the growth habit file.}

\item{recorded.species.codes}{Vector. Species recorded so that generic.growth.habit() can identify unknown codes.}

\item{species.code}{Character. The field name for the species codes in the species file.}

\item{species.duration}{Character. the field name for the Duration field in the species file.}
}
\description{
Gather species data
}
