% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_all}
\alias{gather_all}
\title{Gather data for plot metadata, line-point intercept, gap, vegetation height, soil stability, species inventory, and rangeland health.}
\usage{
gather_all(
  dsn = NULL,
  source,
  data_list = NULL,
  data_types = c("header", "lpi", "height", "gap", "soilstability", "species",
    "rangelandhealth"),
  output_filepath = NULL,
  output_filetype = c("csv", "rds"),
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{source}{Character string. This specifies the expected data format(s)
and determines which gather function will be used. It must be
one or more of of \code{"AIM"}, \code{"TERRADAT"}, \code{"LMF"}, or \code{"DIMA"}.
This is case-insensitive.}

\item{data_list}{Named list of data frames or character strings. These may be
data frames containing the data that will be gathered, character strings
specifying the filepath to CSV or Rdata files containing the data, or character
strings specifying the name of the corresponding feature class in a geodatabase
specified by the argument \code{dsn}. The names in this list must match the
argument names expected by the gather functions, e.g.
\code{"PINTERCEPT"} or \code{"tblPlots"}. See the documentation for those
functions for additional details. If \code{NULL} this will be ignored.
Defaults to \code{NULL}.}

\item{data_types}{Vector of character strings. These are the data types to
gather into long/tall-format. Valid values are \code{"header"}, \code{"lpi"}, \code{"height"}, \code{"gap"}, \code{"soilstability"}, \code{"species"}, and \code{"rangelandhealth"}.
This is case-insensitive. Defaults to \code{c("header", "lpi", "height", "gap", "soilstability", "species", "rangelandhealth")}.}

\item{output_filepath}{Optional character string. The filepath to a folder to
save all outputs to as RDS or CSV files. This folder must already exist. If
this is not \code{NULL} then the function will write the output and will NOT
return any outputs in the R environment. If this is \code{NULL} then no
data will be written to file and the function will return a list of data frames,
one for each data type. Defaults to \code{NULL}.}

\item{output_filetype}{Optional vector of character strings. This determines
what file(s) type to write the outputs as in \code{output_filepath}. Valid file
values are \code{"CSV"} and \code{"RDS"}, case-insensitive. Only considered if
\code{output_filepath} is not \code{NULL}. Defaults to \code{c("csv", "rds")}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
If \code{output_filepath} is \code{NULL}, a list of tall data frames containing reformatted input data. Otherwise, no outputs in the R environment but files written into \code{output_filepath}.
}
\description{
Given wide format AIM/LMF data, produce long/tall-format versions.

This uses the family of gather functions in terradactyl. For additional
documentation regarding those, please see \code{\link[terradactyl:gather_header]{gather_header()}},
\code{\link[terradactyl:gather_lpi]{gather_lpi()}}, \code{\link[terradactyl:gather_height]{gather_height()}},
\code{\link[terradactyl:gather_gap]{gather_gap()}}, \code{\link[terradactyl:gather_soil_stability]{gather_soil_stability()}},
\code{\link[terradactyl:gather_species_inventory]{gather_species_inventory()}},
\code{\link[terradactyl:gather_rangeland_health]{gather_rangeland_health()}}
}
\examples{
# To process all AIM and LMF data in a geodatabase and save them in the
# folder called "output" in the working directory as both CSV and RDS files.
gather_all(dsn = "Path/To/AIM-LMF_Geodatabase.gdb",
           output_filepath = "output")

# To process line-point intercept and gap data from AIM in a geodatabase and
# store them in the environment as a list of data frames with diagnostic
# messages active.
long_data_list <- gather_all(dsn = "Path/To/AIM-LMF_Geodatabase.gdb",
                             source = "aim",
                             data_types = c("lpi",
                                            "gap"),
                             verbose = TRUE)

# To process wide-format LMF height and species inventory data that are
# already in a data frame in the environment.
long_data_list <- gather_all(source = "lmf",
                             data_types = c("height",
                                            "speciesinventory"),
                             data_list = list(PASTUREHEIGHTS = pastureheights_data_frame,
                                              PLANTCENSUS = plantcensus_data_frame))
}
