% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_gap_lmf}
\alias{gather_gap_lmf}
\title{Convert LMF-format gap data into a long/tall format.}
\usage{
gather_gap_lmf(
  dsn = NULL,
  file_type = NULL,
  GINTERCEPT = NULL,
  POINT = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{file_type}{Deprecated. This argument is no longer functional or necessary and is kept for backwards compatibility with legacy code.}

\item{GINTERCEPT}{Optional data frame or character string. If provided, this must contain the
expected gap intercept data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called GINTERCEPT (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{POINT}{Optional data frame or character string. If provided, this must contain the expected
point metadata. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called POINT (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
Convert LMF-format gap data into a long/tall format.
}
\examples{
# Using a geodatabase that contains the tables tblGapHeader and tblGapDetail
gather_gap_lmf(dsn = "data_folder/aim_data.gdb")

# Using CSV files for tblGapHeader and tblGapDetail.
# These two arguments can also take data frames, if the data have already
# been read in as data frames.
gather_gap_lmf(tblGapHeader = "data_folder/gap_headers.csv",
                    tblGapDetail = "data_folder/gap_detail_records.csv")

# Using data frames for tblGapHeader and tblGapDetail
aim_gapheader <- read.csv("data_folder/gap_headers.csv")
aim_gapdetail <- read.csv("data_folder/gap_detail_records.csv")
gather_gap_lmf(tblGapHeader = aim_gapheader,
                    tblGapDetail = aim_gapdetail)
}
