% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_gap_terradat}
\alias{gather_gap_terradat}
\title{Convert AIM-format gap data into a long/tall format.}
\usage{
gather_gap_terradat(
  dsn = NULL,
  tblGapDetail = NULL,
  tblGapHeader = NULL,
  auto_qc_warnings = TRUE,
  drop_na = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. If provided, this must be the filepath
to a geodatabase which contains the feature classes tblGapHeader and
tblGapDetail, e.g. \code{"C:/DATA/AIM.GDB"}. If this is \code{NULL}, then
the arguments \code{tblGapHeader} and \code{tblGapDetail} are required. Defaults to
\code{NULL}.}

\item{tblGapDetail}{Optional data frame. If provided, this must contain the
expected gap data. If \code{NULL} then the argument
\code{dsn} must be provided. Defaults to \code{NULL}.}

\item{tblGapHeader}{Optional data frame. If provided, this must contain the
expected metadata for the gap data. If \code{NULL} then the argument \code{dsn} must be
provided. Defaults to \code{NULL}.}

\item{auto_qc_warnings}{Logical. If \code{TRUE} the function will test the
data and metadata for duplicated records (records which are identical to
each other across all critical variables) and orphaned records (records
which appear in the data but do not have any corresponding metadata
records), either of which can result in erroneous or unexpected outputs.
The number of duplicated and orphaned records will be reported but without
making any changes. Defaults to \code{TRUE}.}

\item{drop_na}{Logical. If \code{TRUE} then records with \code{NA} values in
their Gap variable will be dropped. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
Convert AIM-format gap data into a long/tall format.
}
\examples{
# Using a geodatabase that contains the tables tblGapHeader and tblGapDetail
gather_gap_terradat(dsn = "data_folder/aim_data.gdb")

# Using CSV files for tblGapHeader and tblGapDetail.
# These two arguments can also take data frames, if the data have already
# been read in as data frames.
gather_gap_terradat(tblGapHeader = "data_folder/gap_headers.csv",
                    tblGapDetail = "data_folder/gap_detail_records.csv")

# Using data frames for tblGapHeader and tblGapDetail
aim_gapheader <- read.csv("data_folder/gap_headers.csv")
aim_gapdetail <- read.csv("data_folder/gap_detail_records.csv")
gather_gap_terradat(tblGapHeader = aim_gapheader,
                    tblGapDetail = aim_gapdetail)



}
