% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_header_lmf}
\alias{gather_header_lmf}
\title{Gather AIM plot-level header data}
\usage{
gather_header_lmf(dsn, ..., verbose = FALSE)
}
\arguments{
\item{dsn}{Character string. This must be the filepath to a geodatabase which
contains the relevant feature classes, e.g. \code{"C:/DATA/AIM.GDB"}. The
required feature classes are: POINT, POINTCOORDINATES, COUNTYNM, STATENM,
GPS, ESFSG.}

\item{...}{Additional optional filtering expressions to be used with
\code{dplyr::filter()} to restrict the processing and output to only a
subset of input data by filtering POINT.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A long-format data frame of header data.
}
\description{
This reads in metadata from AIM sampling used as headers for
various methods and returns it as a long-format data frame suitable for use
in indicator calculations with the package \code{terradactyl}. The required
feature classes in the geodatabase specified with \code{dsn} are POINT,
POINTCOORDINATES, GPS, COUNTYNM, STATENM, and ESFSG. The expected formats for
the input data are those used in the Landscape Monitoring Framework.
}
\examples{
# Basic use.
gather_header_lmf(dsn = "data_path/lmf_data.gdb")

# Using a filtering expression to restrict the processing and output to only
# records where the SpeciesState variable in POINT contained either of the
# values "CO" and "NM".
gather_header_lmf(dsn = "data_path/lmf_data.gdb",
                  SpeciesState \%in\% c("CO", "NM"))

}
