% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_header_terradat}
\alias{gather_header_terradat}
\title{Gather AIM plot-level header data}
\usage{
gather_header_terradat(
  dsn = NULL,
  tblPlots = NULL,
  date_tables = NULL,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{tblPlots}{Data frame or character string. If a data frame, must contain the variables PrimaryKey, SpeciesState, PlotID, PlotKey, EcolSite, Latitude, Longitude, State, Elevation, CountyName, EstablishDate, DateLoadedInDb, and SamplingApproach. If a character string, must either correspond to the filepath to a CSV or RDATA file containing a table with those variables or the name of the feature class in the geodatabase provided as \code{dsn} with those variables. If \code{NULL}, the function will attempt to find a feature class called tblPlots (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{date_tables}{Optional (contingent on other arguments) character vector
or list of data frames. This specifies the tables to extract date
information from. If \code{dsn} is not \code{NULL} AND the geodatabase
contains at least one feature class from the set tblLPIHeader,
tblGapHeader, and tblSpecRichHeader, this argument is optional and can be
left as \code{NULL}. Otherwise, if \code{dsn} is not \code{NULL} and the
desired tables to use are named anything else, this must be a vector of
character strings specifying the names of the relevant feature classes. If
\code{tblPlots} is being used, then this must be a list of data frames
containing the relevant data. Defaults to \code{NULL}.}

\item{...}{Additional optional filtering expressions passed to
\code{\link[dplyr:filter]{dplyr::filter()}} to restrict the processing and output to only a
subset of input data.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A long-format data frame of header data.
}
\description{
This reads in metadata from AIM sampling used as headers for
various methods and returns it as a tall/long-format data frame suitable for use
in indicator calculations with the package \code{terradactyl}. Dates are
added from one or more other tables associated with the specific data
collection methods. The expected format is that used in the Terrestrial AIM
Database (TerrADat).
}
\examples{
# Using a geodatabase which already contains the tables tblPLots and at least
# one automatically-recognized table to pull dates from.
gather_header_terradat(dsn = "data_folder/aim_data.gdb")

# Using a data frame containing the contents of a tblPlots.
# In this example, no path to a geodatabase is specified with dsn, so a list
# containing at least one date table must also be provided.
gather_header_terradat(tblPlots = tblPlots_data_frame,
                       date_tables = list(tblLPIHeader,
                       tblGapHeader))

# Using a geodatabase which already contains the tables tblPLots in
# conjunction with a filtering expression to limit the output to records
# where the State variable contained either of the values "CO" and "NM".
gather_header_terradat(dsn = "data_folder/aim_data.gdb",
                       State \%in\% c("CO", "NM"))

}
