% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_height}
\alias{gather_height}
\title{export gather_height_survey123
rdname gather_height
Convert height data into a tall, tidy data frame}
\usage{
gather_height(
  dsn = NULL,
  file_type = "gdb",
  source,
  tblLPIDetail = NULL,
  tblLPIHeader = NULL,
  PASTUREHEIGHTS = NULL,
  autoQC = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. If provided, this must be the filepath
to a geodatabase which contains the feature classes tblLPIHeader and
tblLPIDetail or the feature class PASTUREHEIGHTS, e.g. \code{"C:/DATA/AIM.GDB"}. If this is \code{NULL} and \code{source} is one of \code{"AIM"}, \code{"TerrADat"}, or \code{"DIMA"} (case-insensitive), then
the arguments \code{tblLPIHeader} and \code{tblLPIDetail}. If this is \code{NULL} and \code{source} is \code{"LMF"} (case-insensitive), then the argument \code{PASTUREHEIGHTS} must be provided.
Defaults to \code{NULL}.}

\item{file_type}{Optional character string. This must be \code{"gdb"},
\code{"txt"}, or \code{"csv"} and must match the file type of \code{dsn}.
If \code{dsn} is \code{NULL} this argument is ignored. Only used if \code{source} is \code{"LMF"}. Defaults to
\code{"gdb"}.}

\item{source}{Character string. The data source format. Must be one of \code{"AIM"}, \code{"TerrADat"}, \code{"DIMA"}, or \code{"LMF"} (case independent).}

\item{tblLPIDetail}{Optional data frame. If provided, this must contain the
expected metadata for the LPI data. If \code{NULL} then the argument
\code{dsn} must be provided. Only used if \code{source} is one of \code{"AIM"}, \code{"TerrADat"}, or \code{"DIMA"}. Defaults to \code{NULL}.}

\item{tblLPIHeader}{Optional data frame. If provided, this must contain the
expected LPI data. If \code{NULL} then the argument \code{dsn} must be
provided. Only used if \code{source} is one of \code{"AIM"}, \code{"TerrADat"}, or \code{"DIMA"}. Defaults to \code{NULL}.}

\item{PASTUREHEIGHTS}{Optional data frame. If provided, this must contain the
expected heights data. If \code{NULL} then the argument \code{dsn} must be
provided. Only used if \code{source} is \code{"LMF"}. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}

\item{auto_qc}{Logical (disabled). If \code{TRUE} then AIM/DIMA/TerrADat data will be
checked for non-unique and orphaned records before doing processing. If any
are found, a warning will be triggered but the gather will still be carried
out. It is strongly recommended that any identified issues be addressed to
avoid incorrect records in the output. Defaults to \code{TRUE}.}
}
\value{
A tall data frame containing the data from the height measurements.
}
\description{
Given wide format line-point intercept data, create a tall
format data frame usable by other terradactyl functions.
}
\examples{
gather_height(dsn = "Path/To/AIM_Geodatabase.gdb",
              source = "AIM")
gather_height(dsn = "Path/To/LMF_Geodatabase.gdb",
              source = "LMF")

aim_lpidetail <- read.csv("Path/To/tblLPIDetail.csv")
aim_lpiheader <- read.csv("Path/To/tblLPIHeader.csv")
gather_height(source = "AIM",
              tblLPIDetail = aim_lpidetail,
              tblLPIHeader = aim_lpiheader)

lmf_heights <- read.csv("Path/To/PASTUREHEIGHTS.csv")
gather_height(source = "LMF",
              PASTUREHEIGHTS = lmf_heights)
}
