% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_height_lmf}
\alias{gather_height_lmf}
\title{Convert LMF-format height data into a long format.}
\usage{
gather_height_lmf(
  dsn = NULL,
  file_type = "gdb",
  PASTUREHEIGHTS = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. If provided, this must be
the filepath to a geodatabase which contains the feature class
PASTUREHEIGHTS, e.g. \code{"C:/DATA/LMF.GDB"}. If this is \code{NULL}, then
the argument \code{PASTUREHEIGHTS} must be provided. Defaults to
\code{NULL}.}

\item{file_type}{Optional character string. This must be \code{"gdb"},
\code{"txt"}, or \code{"csv"} and must match the file type of \code{dsn}.
If \code{dsn} is \code{NULL} this argument is ignored. Defaults to
\code{"gdb"}.}

\item{PASTUREHEIGHTS}{Optional data frame. If provided, this must contain the
expected heights data. If \code{NULL} then the argument \code{dsn} must be
provided. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
This reads in height data from LMF sampling and returns it as a long-format data frame suitable for use
in indicator calculations with the package \code{terradactyl}.The expected format is that used in the Landscape Monitoring Framework (LMF).
}
\examples{
# Using a geodatabase that contains the tables tblLPIHeader and tblLPIDetail
gather_height_lmf(dsn = "data_folder/lmf_data.gdb")

# Using CSV files for PASTUREHEIGHTS.
# This argument can also take data frames, if the data have already been read
# in as a data frame.
gather_height_lmf(PASTUREHEIGHTS = "data_folder/lmf_pastureheights.csv")

# Using data frames for PASTUREHEIGHTS
lmf_pastureheights <- read.csv("data_folder/lmf_pastureheights.csv")
gather_height_lmf(PASTUREHEIGHTS = lmf_pastureheights)


}
