% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_height_terradat}
\alias{gather_height_terradat}
\title{Convert AIM-format height data into a long format.}
\usage{
gather_height_terradat(
  dsn = NULL,
  tblLPIDetail = NULL,
  tblLPIHeader = NULL,
  auto_qc_warnings = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. If provided, this must be
the filepath to a geodatabase which contains the feature classes
tblLPIHeader and tblLPIDetail, e.g. \code{"C:/DATA/AIM.GDB"}. If this is
\code{NULL}, then the arguments \code{tblLPIHeader} and \code{tblLPIDetail}
must be provided. Defaults to \code{NULL}.}

\item{tblLPIDetail}{Optional data frame. If provided, this must contain the
expected LPI data. If \code{NULL} then the argument
\code{dsn} must be provided. Defaults to \code{NULL}.}

\item{tblLPIHeader}{Optional data frame. If provided, this must contain the
expected metadata for the LPI data. If \code{NULL} then the argument \code{dsn} must be
provided. Defaults to \code{NULL}.}

\item{auto_qc_warnings}{Logical. If \code{TRUE} the function will test the
data and metadata for duplicated records (records which are identical to
each other across all critical variables) and orphaned records (records
which appear in the data but do not have any corresponding metadata
records), either of which can result in erroneous or unexpected outputs.
The number of duplicated and orphaned records will be reported but without
making any changes. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
This reads in height data from AIM sampling and returns it as a long-format data frame suitable for use
in indicator calculations with the package \code{terradactyl}.The expected format is that used in the Terrestrial AIM
Database (TerrADat).
}
\examples{
# Using a geodatabase that contains the tables tblLPIHeader and tblLPIDetail
gather_height_terradat(dsn = "data_folder/aim_data.gdb")

# Using CSV files for tblLPIHeader and tblLPIDetail.
# These two arguments can also take data frames, if the data have already
# been read in as data frames.
gather_height_terradat(tblLPIHeader = "data_folder/lpi_headers.csv",
                       tblLPIDetail = "data_folder/lpi_detail_records.csv")

# Using data frames for tblLPIHeader and tblLPIDetail
aim_lpiheader <- read.csv("data_folder/lpi_headers.csv")
aim_lpidetail <- read.csv("data_folder/lpi_detail_records.csv")
gather_height_terradat(tblLPIHeader = aim_lpiheader,
                       tblLPIDetail = aim_lpidetail)

}
