% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_lpi}
\alias{gather_lpi}
\title{Function to convert LPI data into a long format.}
\usage{
gather_lpi(
  dsn = NULL,
  file_type = "gdb",
  source,
  tblLPIDetail = NULL,
  tblLPIHeader = NULL,
  PINTERCEPT = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{file_type}{Deprecated. This argument is no longer functional or necessary and is kept for backwards compatibility with legacy code.}

\item{source}{Character string. This specifies the expected data format(s)
and determines which specialized gather function will be used. It must be
one of \code{"AIM"}, \code{"TERRADAT"}, \code{"LMF"}, \code{"DIMA"} or
\code{"NRI"}. This is case-insensitive.}

\item{tblLPIDetail}{Optional data frame or character string. If provided, this must contain the
expected LPI data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called tblLPIDetail (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{tblLPIHeader}{Optional data frame or character string. If provided, this must contain the
expected metadata for the LPI data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called tblLPIHeader (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{PINTERCEPT}{Optional data frame or character string. If provided, this must contain the
expected LPI data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called PINTERCEPT (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
This is a wrapper function for specialized functions which read
in Line-Point Intercept data from AIM anbd LMF sampling and return
them as a long-format data frame suitable for use in indicator calculations
with the package \code{terradactyl}. Supported formats include those used in
AIM (Assessment, Inventory, and Monitoring) and TerrADat (Terrestrial AIM
Database), LMF (Landscape Monitoring Framework), and DIMA (the Database for
Inventory, Monitoring, and Assessment). For additional information about
arguments, see the documentation for the functions \code{\link[terradactyl:gather_lpi_terradat]{gather_lpi_terradat()}}, \code{\link[terradactyl:gather_lpi_lmf]{gather_lpi_lmf()}}, and
\code{\link[terradactyl:gather_lpi_nri]{gather_lpi_nri()}}.
}
\details{
The \code{source} argument determines which other arguments are used or ignored.

When \code{source} is one of \code{"AIM"}, \code{"TERRADAT"}, or \code{"DIMA"}
then the arguments \code{tblLPIHeader} and \code{tblLPIDetail} are both considered.

When \code{source} is \code{"LMF"} then the argument \code{PINTERCEPT} is considered.

Regardless of the value of \code{source}, the data sources represented by those other arguments are required. The simplest way to provide them is to provide the filepath to a geodatabase as \code{dsn} with each of those feature classes appearing by the same name as the corresponding argument in that geodatabase.
}
\examples{
# LPI data from a geodatabase in the format of the Terrestrial AIM Database
gather_lpi(dsn = "data_path/aim_data.gdb,
           source = "terradat")

# LPI data from a geodatabase in the format of the Landscape Monitoring Framework
gather_lpi(dsn = "data_path/lmf_data.gdb,
           source = "lmf")


}
