% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_lpi_lmf}
\alias{gather_lpi_lmf}
\title{Function to convert LMF-format LPI data into a long format.}
\usage{
gather_lpi_lmf(
  dsn = NULL,
  file_type = "gdb",
  PINTERCEPT = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. This is the filepath to
the source for the LMF data. If this points to a geodatabase, the
geodatabase must contain the feature class PINTERCEPT with the LPI data and
\code{file_type} must be \code{"gdb"}. If this points to a TXT or CSV file,
\code{file_type} must be \code{"txt"} and \code{"csv"}, respectively. If
\code{NULL}, then a data frame must be provided as the argument
\code{PINTERCEPT}. Defaults to \code{NULL}.}

\item{file_type}{Optional character string. This must be \code{"gdb"},
\code{"txt"}, or \code{"csv"} and must match the file type of \code{dsn}.
If \code{dsn} is \code{NULL} this argument is ignored. Defaults to
\code{"gdb"}.}

\item{PINTERCEPT}{Optional (contingent) data frame. This is the data frame
containing the LPI data in the format of the PINTERCEPT table used by LMF.
If provided, this will be used even if \code{dsn} is also non-NULL.
Defaults to \code{NULL}.}

\item{Logical.}{If \code{TRUE} then the function will report back diagnostic
information as console messages while it works. Defaults to \code{FALSE}.}
}
\description{
This reads in Line-Point Intercept from LMF sampling and returns it as a long-format data frame suitable for use
in indicator calculations with the package \code{terradactyl}.The expected format is that used in the Landscape Monitoring Framework (LMF).
}
\examples{
# Using a geodatabase that contains the table PINTERCEPT.
gather_lpi_lmf(dsn = "data_folder/lmf_data.gdb")

# Using a CSV file for PINTERCEPT.
# This argument can also take data frames, if the data have already been read
# in as a data frame.
gather_lpi_lmf(PINTERCEPT = "data_folder/lpi_detail_records.csv")


}
