% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_rangeland_health_lmf}
\alias{gather_rangeland_health_lmf}
\title{Convert LMF-format Interpreting Indicators of RangelandHealth data into a long/tall format.}
\usage{
gather_rangeland_health_lmf(
  dsn = NULL,
  file_type = NULL,
  RANGEHEALTH = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. If provided, this must be the filepath
to a geodatabase which contains the feature class RANGEHEALTH, e.g. \code{"C:/DATA/AIM.GDB"}. If this is \code{NULL}, then
the argument \code{RANGEHEALTH} is required. Defaults to
\code{NULL}.}

\item{file_type}{Character string. The file extension for the path in \code{dsn}.}

\item{RANGEHEALTH}{Optional data frame. If provided, this must contain the
expected IIRH data. If \code{NULL} then the argument
\code{dsn} must be provided. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
Convert LMF-format Interpreting Indicators of RangelandHealth data into a long/tall format.
}
\examples{
# Using a geodatabase that contains the table RANGEHEALTH.
gather_rangeland_health_lmf(dsn = "data_folder/lmf_data.gdb")

# Using a data frame for RANGEHEALTH.
lmf_iirh <- read.csv("data_folder/RANGEHEALTH_detail_records.csv")
gather_rangeland_health_lmf(RANGEHEALTH = lmf_iirh)

}
