% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_soilstability.R
\name{gather_soil_stability_terradat}
\alias{gather_soil_stability_terradat}
\alias{gather_soil_stability_lmf}
\alias{gather_soil_stability}
\title{Gather Soil Stability Data}
\usage{
gather_soil_stability_terradat(dsn)

gather_soil_stability_lmf(dsn, file_type = "gdb")

gather_soil_stability(dsn, source, file_type = "gdb")
}
\arguments{
\item{dsn}{Character string. Full filepath including file extension to the
geodatabase that contains soil stability data.}

\item{source}{String. Specifies the original data source.}
}
\description{
Gather Soil Stability Data
}
