% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_soil_summary.R
\name{gather_soil_summary}
\alias{gather_soil_summary}
\alias{gather_soil_summary_lmf}
\alias{gather_soil_summary_terradat}
\title{Gather TerrADat and LMF Soil data into a summary data frame}
\usage{
gather_soil_summary_lmf(dsn = NULL, SOILHORIZON = NULL)

gather_soil_summary_terradat(
  dsn = NULL,
  tblSoilPitHorizons = NULL,
  tblSoilPits = NULL
)

gather_soil_summary(
  dsn = NULL,
  source,
  tblSoilPitHorizons = NULL,
  tblSoilPits = NULL,
  SOILHORIZON = NULL
)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file 
extension) of the geodatabase containing the table of interest. This field 
is unnecessary if you supply either both of tblSoilPitHorizons and 
tblSoilPits (AIM/DIMA/TerrADat) or SOILHORIZON (LMF/NRI).}

\item{SOILHORIZON}{Dataframe of the data structure SOILHORIZON from LMF/NRI 
database with the addition of PrimaryKey and DBKey fields. Use when data 
source is LMF or NRI; alternately provide dsn.}

\item{tblSoilPitHorizons}{Dataframe of the data structure tblSoilPitHorizons 
from the DIMA database with the addition of PrimaryKey and DBKey fields. 
Use when data source is AIM, DIMA, or TerrADat; alternately provide dsn.}

\item{tblSoilPits}{Dataframe of the data structure tblSoilPits from the DIMA 
database with the addition of PrimaryKey and DBKey fields. Use when data 
source is AIM, DIMA, or TerrADat; alternately provide dsn.}

\item{source}{Character string. The data source format, 
\code{"AIM", "TerrADat", "DIMA", "LMF", "NRI"} (case independent).}
}
\value{
A tall data frame summarizing horizon data to the soil pit
}
\description{
Given soil horizon and pit data, create a tall format data frame
usable by other terradactyl functions.
}
\examples{
gather_soil_summary(dsn = "Path/To/AIM_Geodatabase.gdb", 
                    source = "AIM")
gather_soil_summary(dsn = "Path/To/LMF_Geodatabase.gdb", 
                    source = "LMF")

aim_horizons <- read.csv("Path/To/tblSoilPitHorizons.csv")
aim_pits <- read.csv("Path/To/tblSoilPits.csv")
gather_soil_summary(source = "AIM", 
                    tblSoilPitHorizons = aim_horizons,
                    tblSoilPits = aim_pits)

lmf_horizons <- read.csv("Path/To/SOILHORIZON.csv")
gather_soil_summary(source = "LMF", 
                    SOILHORIZON = lmf_horizons)
}
\seealso{
Other <gather>: 
\code{\link{gather_all}()},
\code{\link{gather_gap}()},
\code{\link{gather_height}()},
\code{\link{gather_lpi}()},
\code{\link{gather_plot_characterization}()},
\code{\link{gather_rangeland_health}()},
\code{\link{gather_soil_horizon}()},
\code{\link{gather_soil_stability}()},
\code{\link{gather_species_inventory}()}
}
\concept{<gather>}
