% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_soilstability.R
\name{gather.soil.stability.terradat}
\alias{gather.soil.stability.terradat}
\alias{gather.soil.stability.lmf}
\alias{gather.soil.stability}
\title{Gather Soil Stability Data}
\usage{
gather.soil.stability.terradat(dsn)

gather.soil.stability.lmf(dsn, file.type = "gdb")

gather.soil.stability(dsn, source, file.type = "gdb")
}
\arguments{
\item{dsn}{Character string. Full filepath including file extension to the geodatabase that contains the tables tblSoilStabDetail and tblSoilStabHeader.}
}
\description{
Gather Soil Stability Data
}
