% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_species_inventory}
\alias{gather_species_inventory}
\title{export gather_species_inventory_survey123
rdname gather_species_inventory
Species Inventory Gather wrapper}
\usage{
gather_species_inventory(
  dsn = NULL,
  source,
  tblSpecRichDetail = NULL,
  tblSpecRichHeader = NULL,
  PLANTCENSUS = NULL,
  file_type = "gdb",
  autoQC = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{source}{Character string. The data source format. Must be one of \code{"AIM"}, \code{"TerrADat"}, \code{"DIMA"}, or \code{"LMF"} (case independent).}

\item{tblSpecRichDetail}{Optional data frame or character string. If provided, this must contain the
expected species inventory data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called tblSpecRichDetail (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{tblSpecRichHeader}{Optional data frame or character string. If provided, this must contain the
expected metadata for the species inventory data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called tblSpecRichHeader (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{PLANTCENSUS}{Optional data frame or character string. If provided, this must contain the
expected species data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called PLANTCENSUS (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{file_type}{Deprecated. This argument is no longer functional or necessary and is kept for backwards compatibility with legacy code.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\value{
A tall data frame containing the data from the species inventory inputs.
}
\description{
Given wide format species richness data, create a tall
format data frame usable by other terradactyl functions. For additional information about
arguments, see the documentation for the functions \code{\link[terradactyl:gather_species_inventory_terradat]{gather_species_inventory_terradat()}} and \code{\link[terradactyl:gather_species_inventory_lmf]{gather_species_inventory_lmf()}}.
}
\details{
The \code{source} argument determines which other arguments are used or ignored.

When \code{source} is one of \code{"AIM"}, \code{"TERRADAT"}, or \code{"DIMA"}
then the arguments \code{tblSpecRichHeader} and \code{tblSpecRichDetail} are both considered.

When \code{source} is \code{"LMF"} then the argument \code{PLANTCENSUS} is considered.

Regardless of the value of \code{source}, the data sources represented by those other arguments are required. The simplest way to provide them is to provide the filepath to a geodatabase as \code{dsn} with each of those feature classes appearing by the same name as the corresponding argument in that geodatabase.
}
