% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_species_inventory_lmf}
\alias{gather_species_inventory_lmf}
\title{Convert LMF-format soil stability data into a long/tall format}
\usage{
gather_species_inventory_lmf(
  dsn = NULL,
  file_type = "gdb",
  PLANTCENSUS = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{file_type}{Deprecated. This argument is no longer functional or necessary and is kept for backwards compatibility with legacy code.}

\item{PLANTCENSUS}{Optional data frame or character string. If provided, this must contain the
expected species data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called PLANTCENSUS (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
Convert LMF-format soil stability data into a long/tall format
}
\examples{
# Using a geodatabase that contains the table SOILDISAG.
gather_species_inventory_lmf(dsn = "data_folder/lmf_data.gdb")

# Using a data frame for SOILDISAG.
lmf_plantcensus <- read.csv("data_folder/PLANTCENSUS_detail_records.csv")
gather_species_inventory_lmf(PLANTCENSUS = lmf_plantcensus)

}
