% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_species_inventory_terradat}
\alias{gather_species_inventory_terradat}
\title{Convert AIM-format species inventory data into a long/tall format}
\usage{
gather_species_inventory_terradat(
  dsn = NULL,
  tblSpecRichDetail = NULL,
  tblSpecRichHeader = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{tblSpecRichDetail}{Optional data frame or character string. If provided, this must contain the
expected species inventory data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called tblSpecRichDetail (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{tblSpecRichHeader}{Optional data frame or character string. If provided, this must contain the
expected metadata for the species inventory data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called tblSpecRichHeader (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
Convert AIM-format species inventory data into a long/tall format
}
