% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.nri.R
\name{generate_pkey_nri}
\alias{generate_pkey_nri}
\title{Generate PrimaryKeys to nri data}
\usage{
generate_pkey_nri(POINTCOORDINATES, sensitive_data_path)
}
\arguments{
\item{POINTCOORDINATES}{as a data frame, the POINTCOORDINATES data}

\item{sensitive_data_path}{path where UID data frame with sensitive and UID PrimaryKeys is stored}
}
\value{
UID file written to sensitive data path
}
\description{
Assign PrimaryKeys to every unique point in NRI data
}
