% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{generic_growth_habits}
\alias{generic_growth_habits}
\title{Attribute growth habits and durations to generic species codes}
\usage{
generic_growth_habits(
  data,
  data_code = "code",
  species_list,
  species_code = "SpeciesCode",
  species_growthhabit = "GrowthHabit",
  species_growthhabitsub = "GrowthHabitSub",
  species_duration = "Duration",
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame. The source data which contains species codes in the variable specified by \code{species_code}. The only required variable is the one containing the species codes; all others will be ignored.}

\item{data_code}{Character string. The name of the variable in \code{data} which contains the species codes. Defaults to \code{"code"}.}

\item{species_list}{Data frame. The existing species characteristics information containing the variables specified with \code{species_code}, \code{species_growthhabit}, \code{species_growthhabitsub}, and \code{species_duration}. For Terrestrial AIM, this is the output from \code{species_read_aim)()}.}

\item{species_code}{Character string. The name of the variable in \code{species_list} which contains the species codes. Defaults to \code{"SpeciesCode"}.}

\item{species_growthhabit}{Character string. The name of the variable in \code{species_list} which contains the general growth habits, e.g. \code{"woody"}. Defaults to \code{"GrowthHabit"}.}

\item{species_growthhabitsub}{Character string. The name of the variable in \code{species_list} which contains the specific growth habits, e.g. \code{"graminoid"}. Defaults to \code{"GrowthHabitSub"}.}

\item{species_duration}{Character string. The name of the variable in \code{species_list} which contains the species duration, e.g. \code{"annual"} or \code{"perennial"}. Defaults to \code{"Duration"}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
The original data frame provided as \code{species_list} with records added for identified generic codes found in \code{data}.
}
\description{
Given a data set with species codes and a species list containing growth habit and duration assignments,
this function will identify species codes present in the data which do not occur in the species list
and assign growth habits and durations to those that can be identified as standardized unknown species codes.

Unknown species code formats supported include Terrestrial AIM and LMF. The identification is done using regular expressions and so will only find generic species codes which follow the exact patterns used by AIM and LMF.

The output is the original species list data frame with any successfully identified and attributed
generic species codes appended.
}
