% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R
\name{height_calc}
\alias{height_calc}
\title{Calculate the standard Terrestrial AIM Database (TerrADat) height indicators}
\usage{
height_calc(
  header,
  height_tall,
  species_file = species_file,
  source = NULL,
  generic_species_file = NULL,
  digits = 1,
  verbose = FALSE
)
}
\arguments{
\item{header}{Optional data frame or character string. The metadata for the plots involved in the calculations, this is used to filter or subset the data being used for the calculations and must contain this must contain the variable PrimaryKey. If a character string, this must point to a CSV file containing the data. If \code{NULL} then no filtering will occur. Defaults to \code{NULL}.}

\item{species_file}{Data frame or character string. The species characteristics information. If this is a character string for the filepath to a geodatabase, that geodatabase must contain both the tblNationalPlants and tblStateSpecies tables. Otherwise, this must either be the output from \code{species_read_aim()} or be a character string pointing to a CSV file containing the output from \code{species_read_aim()}.}

\item{source}{Character string. If \code{"terradat"} or \code{"aim"} (case insensitive) then live and "dead" heights will be calculated. Defaults to \code{NULL}.}

\item{generic_species_file}{Optional character string. Must specify the full path to a CSV containing generic species information. If this is \code{NULL}. Defaults to \code{NULL}.}

\item{digits}{Integer. The number of decimal places that the output values will be rounded to. Values larger than \code{1} are not recommended because they will likely imply false precision. Defaults to \code{1}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}

\item{gap_tall}{Data frame or character string. The long/tall-format gap data for the plots involved in the calculations. The format must match the output from \code{gather_gap()}. If a character string, this must point to a CSV file containing the data.}

\item{species_code_var}{Character string. The name of the variable in the species characteristics that contains the species codes. Defaults to \code{"SpeciesCode"}.}
}
\value{
A data frame matching the format of height indicators in TerrADat.
}
\description{
This function calculates the full set of height-derived indicators that are standard for TerrADat.

For any other gap indicators, use the underlying function \code{gap_cover()}.
}
