% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IIRH.R
\name{gather.rangeland.health.terradat}
\alias{gather.rangeland.health.terradat}
\alias{gather_rangeland_health_lmf}
\alias{gather.rangeland.health}
\title{Rangeland Health}
\usage{
gather.rangeland.health.terradat(dsn)

gather_rangeland_health_lmf(dsn, file.type = "gdb")

gather.rangeland.health(dsn, layer, file.type = "gdb")
}
\arguments{
\item{dsn}{}
}
\description{
Rangeland Health
}
