% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicators_sample}
\alias{indicators_sample}
\title{Terradactyl indicator data}
\description{
Sample of indicator data produced by the build_indicators function, using the sample tall tables included in this package.
}
\keyword{data}
\keyword{indicator}
