% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.nri.R
\name{nri_field_names}
\alias{nri_field_names}
\alias{read_nri_text}
\alias{ingest_nri}
\alias{nri_subset}
\title{Read NRI text files into a gdb}
\usage{
nri_field_names(dsn)

read_nri_text(table_name, dsn, DBKey = "auto")

ingest_nri(
  dsn,
  out,
  dsn_type = "gdb",
  rda = TRUE,
  uid = "",
  pwd = "",
  connection = ""
)

nri_subset(data, PK_subset, out)
}
\arguments{
\item{dsn}{Vector of file paths to the folders where files are held. The last folder in the directory will be used to set the DBKey.}

\item{table_name}{Name of the NRI data table (e.g., \code{POINTCOORDINATES})}

\item{out}{String. Filepath to the output database}

\item{dsn_type}{Type of database. Defaults to geodatabase}

\item{data}{Dateframe. NRI data table}
}
\value{
List of tables and if specified a populated database
}
\description{
Ingestion script to convert NRI text file format to database format
}
