% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R
\name{lpi_calc}
\alias{lpi_calc}
\title{Calculate the standard Terrestrial AIM Database (TerrADat) Line-Point Intercept indicators}
\usage{
lpi_calc(
  header,
  lpi_tall,
  species_file,
  species_code_var = "SpeciesCode",
  indicators_vars = NULL,
  expected_indicator_names,
  apply_species_adjustment = TRUE,
  generic_species_file = NULL,
  digits = 6,
  verbose = FALSE
)
}
\arguments{
\item{header}{Data frame or character string. The metadata for the plots involved in the calculations, this must contain the variable PrimaryKey and any of SpeciesState, State, and County. If a character string, this must point to a CSV file containing the data.}

\item{lpi_tall}{Data frame or character string. The long/tall-format LPI data for the plots involved in the calculations. The format must match the output from \code{gather_lpi()}. If a character string, this must point to a CSV file containing the data.}

\item{species_file}{Data frame or character string. The species characteristics information. If this is a character string for the filepath to a geodatabase, that geodatabase must contain both the tblNationalPlants and tblStateSpecies tables. Otherwise, this must either be the output from \code{species_read_aim()} or be a character string pointing to a CSV file containing the output from \code{species_read_aim()}.}

\item{species_code_var}{Character string. The name of the variable in the species characteristics provided as \code{species_file} that contains the species codes. Defaults to \code{"SpeciesCode"}.}

\item{generic_species_file}{Optional character string. Must specify the full path to a CSV containing generic species information. If this is \code{NULL}. Defaults to \code{NULL}.}

\item{digits}{Integer. The number of decimal places that the output values will be rounded to. Values larger than \code{2} are not recommended because they will likely imply false precision. Defaults to \code{6}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A data frame matching the format of LPI indicators in TerrADat.
}
\description{
This function calculates the full set of LPI-derived indicators that are standard for TerrADat.
These indicators are dependent on the species characteristics used by Terrestrial AIM and stored in the tblNationalPlants and tblStateSpecies tables in TerrADat. Attempting to use this function with any other format of species characteristic data will almost certainly fail.

For any other LPI-derived indicators, use the underlying functions \code{pct_cover()} and \code{mean_height()}.
}
