% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{lpi_indicator_definitions}
\alias{lpi_indicator_definitions}
\title{Definitions for TerrADat indicator source variables}
\usage{
lpi_indicator_definitions()
}
\description{
This function takes no arguments and returns a structured list used by \code{\link[terradactyl:adjust_species_attributes]{adjust_species_attributes()}} to modify and add variables to LPI data with associated species attributes.
}
