% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{mean.height}
\alias{mean.height}
\title{Calculate the vegetation height
Mean or max}
\usage{
\method{mean}{height}(lpi.height.tall, method = "grouped",
  omit.zero = FALSE, by.line = FALSE, ...)
}
\arguments{
\item{method}{Indicatoes the type of indicator, \code{"max"}, which yields the average maximum height (of the herbaceous or woody heights) on the plot or \code{"mean"} which yields the mean height by functional group (woody/herbaceous).}

\item{omit.zero}{Logical. If \code{TRUE} the results omit height measurements of \code {0}}

\item{by.line}{Logical.}

\item{...}{One or more bare variable name from \code{lpi.tall} to calculate percent cover for, e.g. \code{GrowthHabitSub} to calculate percent cover by growth habits or \code{GrowthHabitSub, Duration} to calculate percent cover for categories like perennial forbs, annual graminoids, etc.}

\item{lpi.tall}{A tall/long-format data frame. Use the data frame \code{"height"} from the \code{gather.lpi()} output.}
}
