% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_cover.R
\name{pct_bareground}
\alias{pct_bareground}
\title{Calculate percent bare ground from Line-Point Intercept data}
\usage{
pct_bareground(
  lpi_tall,
  tall = FALSE,
  by_line = FALSE,
  indicator_variable = "code",
  soil_values = c("S"),
  digits = 1,
  verbose = FALSE
)
}
\arguments{
\item{lpi_tall}{Data frame. The long/tall-format line-point intercept data. This must be in the format produced by \code{gather_lpi()}.}

\item{tall}{Logical. If \code{TRUE} then the output will be in a long/tall format. If \code{FALSE} then the output will be wide. Defaults to \code{FALSE}.}

\item{by_line}{Logical. If \code{TRUE} then the cover will be calculated for each unique combination of values in the PrimaryKey and LineKey variables instead of simply by PrimaryKey. Defaults to \code{FALSE}.}

\item{indicator_variable}{Optional character string. This must be the variable name in \code{lpi_tall} containing the value \code{"S"} for records where the pin came to rest on soil. Defaults to \code{"code"}.}

\item{soil_values}{Character string or vector of character strings. The values in the variable specified with \code{indicator_variable} that correspond to uncovered ground. Defaults to \code{c("S")}.}

\item{digits}{Integer. The number of decimal places that the output cover percentages will be rounded to. Values larger than \code{2} are not recommended because they will likely imply false precision. Defaults to \code{1}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A data frame containing PrimaryKeys and a variable called BareSoil containing the percent of bare soil.
}
\description{
Calculate the percent of a transect with bare soil at the surface and no cover from vegetation or rocks.
This is a simple wrapper for \code{pct_cover()} and is equivalent to calculating first hit cover for the value \code{"S"}.
}
