% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_cover.R
\name{pct_cover}
\alias{pct_cover}
\title{Basic percent cover function}
\usage{
pct_cover(
  lpi_tall,
  tall = FALSE,
  hit = "any",
  by_line = FALSE,
  ...,
  indicator_variables = NULL,
  digits = 1,
  verbose = FALSE
)
}
\arguments{
\item{lpi_tall}{The tall LPI data}
}
\description{
Calculate the percent cover
}
