% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.query}
\alias{plot.query}
\title{Select plots using spatial extent}
\usage{
\method{plot}{query}(gathered.data, extent, plots = NULL)
}
\arguments{
\item{gathered.data}{Data frame. The plot data containing at minimum a variable called "PrimaryKey" containing values matching those in \code{extent} or \code{plots}.}

\item{extent}{Spatial points/polygons data frame. Spatial restriction for the results. If \code{plots} is not provided, then this must contain a variable called "PrimaryKey" matching the values in \code{gathered.data}.}

\item{plots}{Optional spatial points data frame. Must have a variable called "PrimaryKey" containing values matching those in \code{gathered.data}. Will be restricted to the intersection with \code{extent}.}
}
\description{
Subset plots using the PrimaryKey values. This can either be using the values that appear in the spatial data frame \code{extent} or in the spatial points data frame \code{plots} restricted to their intersection with \code{extent}.
}
