% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.nri.R
\name{read_nri_text}
\alias{read_nri_text}
\title{Read NRI txt files using table_name}
\usage{
read_nri_text(
  sensitive_data_path,
  table_name,
  dsn,
  DBKey = "auto",
  GL_schema_path
)
}
\arguments{
\item{sensitive_data_path}{path to the folder where the sensitive PrimaryKey and UID PrimaryKey will be stored}

\item{table_name}{Character string of the nri table names}

\item{dsn}{File path where NRI text files stored}

\item{GL_schema_path}{Tables and names from NRI Grazing Land Guide}
}
\value{
NRI files with column name assigned
}
\description{
Assign the column header to the NRI text files and save
}
