% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{read_whatever}
\alias{read_whatever}
\title{Read in various formats smoothly}
\usage{
read_whatever(
  input,
  layer = NULL,
  regex = FALSE,
  best_guess = TRUE,
  accept_failure = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{input}{Data frame or character string. If a character string, this should point to an RDS, Rdata, CSV, or GDB file.}

\item{layer}{Optional character string. If this is not \code{NULL} and \code{input} points to a geodatabase, this is the name of the table or feature class to try to read in. Defaults to \code{NULL}.}

\item{regex}{Logical. If \code{TRUE} then \code{layer} will be treated as a regular expression. Defaults to \code{FALSE}.}

\item{best_guess}{Logical. If \code{regex} is \code{TRUE} and this is \code{TRUE} then in the case that multiple layers in the geodatabase match the regular expression \code{layer} the one with the shortest name will be used. If \code{FALSE} and multiple layers match the regular expression the function will stop and return an error. Defaults to \code{TRUE}.}

\item{accept_failure}{Logical. If \code{FALSE} and \code{regex} is \code{TRUE} then in the case that no layers match the regular expression the function will stop an return an error. If \code{TRUE} then the function will return \code{NULL}. Defaults to \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\description{
This function will flexibly read in a variety or formats including RDS, Rdata, CSV, and geodatabase feature classes.
It will also accept data frames, although those will be returned unaltered.
This exists to facilitate legacy code that was originally written using terradactyl v1.1.0 and earlier.
}
