% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{read_with_fallback}
\alias{read_with_fallback}
\title{Read in data with a fallback option}
\usage{
read_with_fallback(
  dsn = NULL,
  tbl = NULL,
  default_name = NULL,
  regex = FALSE,
  best_guess = FALSE,
  accept_failure = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this should be the filepath to the geodatabase where the feature class specified by \code{default_name} is expected. This is passed on to \code{read_whatever()} as the argument \code{input}. Defaults to \code{NULL}.}

\item{tbl}{Optional character string or data frame. If this is not \code{NULL} and \code{dsn} points to a geodatabase, this will be used as the \code{layer} argument in \code{read_whatever()}. Otherwise, this is passed on to \code{read_whatever()} as the argument \code{input}. Defaults to \code{NULL}.}

\item{regex}{Logical. If \code{TRUE} then \code{tbl} will be treated as a regular expression as necessary. Defaults to \code{FALSE}.}

\item{best_guess}{Logical. If \code{regex} is \code{TRUE} and this is \code{TRUE} then in the case that multiple layers in a geodatabase match the regular expression \code{tbl} the one with the shortest name will be used. If \code{FALSE} and multiple layers match the regular expression the function will stop and return an error. Defaults to \code{TRUE}.}

\item{accept_failure}{Logical. If \code{FALSE} and \code{regex} is \code{TRUE} then in the case that no layers match the regular expression the function will stop an return an error. If \code{TRUE} then the function will return \code{NULL}. Defaults to \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\description{
This is a wrapper for \code{\link[terradactyl:read_whatever]{read_whatever()}} that adds an option for a default.
The primary use case is in situations where there is an expected feature class in a geodatabase that should be used if the user-provided input is unusable or \code{NULL}.
}
\details{
The order of operations is:
\enumerate{
  \item If \code{tbl} is not \code{NULL}, the function will try to figure out how to use it.
    \enumerate{
      \item Check to see if \code{tbl} is a data frame. If so, assign it as output and move on. Otherwise try the next step.
      \item Check to see if \code{tbl} is a character string ending in a file extension. If so, use \code{read_whatever()} to assign it to output and move on. Otherwise try the next step.
      \item Check to see if \code{tbl} is a character string without a file extension. If so AND \code{dsn} is a filepath to a geodatabase, try to use \code{tbl} as a feature class name. Otherwise, throw an error.
    }
  \item If \code{tbl} is \code{NULL}, try to use \code{dsn} with \code{read_whatever()} looking for \code{layer = default_name} with \code{regex = TRUE} and \code{best_guess = TRUE}.
  \item If nothing can be read in, throw an error.
}

Note that \code{accept_failure} doesn't apply to the whole process and is only passed on to any \code{read_whatever()} calls.
}
