% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHEM.R
\name{RHEM}
\alias{RHEM}
\title{RHEM Input Calculations}
\usage{
RHEM(lpi_species, header, slope_shape)
}
\arguments{
\item{lpi_species}{Dataframe of lpi tall table joined with RHEM species attributes in the field "RHEM_Habit"}

\item{slope_shape}{Dataframe of slope tall table with field SlopeShape = Vertical Slope Shape}
}
\description{
RHEM Input Calculations
}
