% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrub.shape.R
\name{sagebrush_shape_base}
\alias{sagebrush_shape_base}
\alias{sagebrush_shape}
\title{Sagebrush Shape}
\usage{
sagebrush_shape_base(lpi_tall)

sagebrush_shape(lpi_tall, live = TRUE)
}
\arguments{
\item{lpi_tall}{Dataframe. Gathered LPI dataframe in a tall format.}

\item{live}{Logical. Indicate if you want to distinguish live vs dead shrub
shapes.Default is \code{TRUE}}
}
\value{
Data frame of count and predominants of different sagebrush shapes
}
\description{
Describe the shape characteristics of sagebrush on the plot,
including number of sagebrush hits in each shape type and
the predominant shape_
}
