% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R
\name{soil_stability_calc}
\alias{soil_stability_calc}
\title{Calculate the standard Terrestrial AIM Database (TerrADat) soil stability indicators}
\usage{
soil_stability_calc(soil_stability_tall, digits = 6, verbose = FALSE)
}
\arguments{
\item{soil_stability_tall}{Data frame or character string. The long/tall-format gap data for the plots involved in the calculations. The format must match the output from \code{gather_soil_stability()}. If a character string, this must point to a CSV file containing the data.}

\item{digits}{Integer. The number of decimal places that the output values will be rounded to. Values larger than \code{1} are not recommended because they will likely imply false precision. Defaults to \code{1}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}

\item{species_file}{Data frame or character string. The species characteristics information. If this is a character string for the filepath to a geodatabase, that geodatabase must contain both the tblNationalPlants and tblStateSpecies tables. Otherwise, this must either be the output from \code{species_read_aim()} or be a character string pointing to a CSV file containing the output from \code{species_read_aim()}.}

\item{species_code_var}{Character string. The name of the variable in the species characteristics that contains the species codes. Defaults to \code{"SpeciesCode"}.}

\item{source}{Character string. If \code{"terradat"} or \code{"aim"} (case insensitive) then live and "dead" heights will be calculated. Defaults to \code{NULL}.}

\item{generic_species_file}{Optional character string. Must specify the full path to a CSV containing generic species information. If this is \code{NULL}. Defaults to \code{NULL}.}
}
\value{
A data frame matching the format of height indicators in TerrADat.
}
\description{
This function calculates the full set of soil stability-derived indicators that are standard for TerrADat.
}
