% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_count}
\alias{species_count}
\title{Get a count of species}
\usage{
species_count(
  species_inventory_tall,
  ...,
  indicator_variables = NULL,
  verbose = FALSE
)
}
\arguments{
\item{species_inventory_tall}{Data frame. The long format species inventory data from \code{gather_spp_inventory()}.}

\item{indicator_variables}{Optional character string or vector of character strings. The names of variables to group by for counts.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\description{
Count the number of unique species found.
}
