% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_inventory.R
\name{gather_species_inventory_terradat}
\alias{gather_species_inventory_terradat}
\alias{species_count}
\alias{tall_species}
\alias{gather_species_lmf}
\alias{gather_species_inventory}
\title{Species Inventory}
\usage{
gather_species_inventory_terradat(dsn)

species_count(species_inventory_tall, ...)

tall_species(species_inventory_detail)

gather_species_lmf(dsn, file_type = "gdb")

gather_species_inventory(dsn, source, file_type = "gdb")
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extension)
of the geodatabase containing the table of interest.}

\item{...}{Grouping variables for species counts}

\item{source}{Character string. Data format source,
\code{"AIM", "TerrADat", "LMF", "NRI"} are all valid options.}
}
\value{
A data frames containing the data from the species inventory data in tall format.
}
\description{
Create a tall species inventory data frame.
}
