% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_QC.R
\name{species_list_check}
\alias{species_list_check}
\title{Check that the species list covers all observed species}
\usage{
species_list_check(dsn_tall, species_list_file, ..., verbose = FALSE)
}
\arguments{
\item{dsn_tall}{The observed data data source}

\item{species_list_file}{The file path to the species list. If the dsn contains the species list, then specify the layer within the dsn.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\description{
Quality control species attribute lists prior to running indicator calculations.
}
