% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_QC.R
\name{species_list_compare}
\alias{species_list_compare}
\title{Find mismatches in species attributions}
\usage{
species_list_compare(species_file, folder, verbose = FALSE)
}
\arguments{
\item{species_file}{Inputs species}

\item{folder}{A folder to write to}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\description{
This finds attributes in species lists which disagree
}
