% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_read_aim}
\alias{species_read_aim}
\title{Read in and combine tblNationalPlants and tblStateSpecies}
\usage{
species_read_aim(dsn, verbose = FALSE)
}
\arguments{
\item{dsn}{Character string. The filepath (including the file extension) to
the geodatabase containing both tblNationalPlants and tblStateSpecies,
e.g., \code{"C:/data/terradat.gdb"}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A data frame with one record per plant code and all expected
  variables formatted appropriately for use with functions like
  \code{lpi_calc()}.
}
\description{
The BLM Terrestrial AIM program uses two separate tables to
store plant species characteristics used to calculate standard indicators for
the Terrestrial AIM Database (TerrADat). Both tables are available through
the AIM data portal and are included with geodatabase exports of TerrADat.
The main table, tblNationalPlants contains information including taxonomic
categories, growth habits, durations, and native and noxious statuses. The
other table, tblStateSpecies, contains per-species assignments to categories
relevant to sage-grouse habitat management. This function reads in both
tables from a geodatabase and combines them, reformatting the contents of
tblStateSpecies to do so. The output format is suitable for use with
indicator calculation functions.

This function is used by functions which calculate the standard set of
TerrADat indicators and can also be used in other workflows in order to use
the specific species characteristics assigned for Terrestrial AIM. These
tables represent an effort to standardize and harmonize species data and
reflect certain decisions, including treating any species known to live
longer than one year as perennial. For additional information about
tblNationalPlants, see the documentation available through the BLM AIM
program.
}
