% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R
\name{spp_inventory_calc}
\alias{spp_inventory_calc}
\title{Calculate the standard Terrestrial AIM Database (TerrADat) species indicators}
\usage{
spp_inventory_calc(
  header,
  spp_inventory_tall,
  species_file,
  species_code_var = "SpeciesCode",
  generic_species_file = NULL,
  source = NULL,
  verbose = FALSE
)
}
\arguments{
\item{header}{Data frame or character string. The metadata for the plots involved in the calculations, this is used to add the SpeciesState variable by joining with the PrimaryKey variable. If a character string, this must point to a CSV file containing the data.}

\item{spp_inventory_tall}{Data frame or character string. The long/tall-format species inventory data for the plots involved in the calculations. The format must match the output from \code{gather_species_inventory()}. If a character string, this must point to a CSV file containing the data.}

\item{species_file}{Data frame or character string. The species characteristics information. If this is a character string for the filepath to a geodatabase, that geodatabase must contain both the tblNationalPlants and tblStateSpecies tables. Otherwise, this must either be the output from \code{species_read_aim()} or be a character string pointing to a CSV file containing the output from \code{species_read_aim()}.}

\item{species_code_var}{Character string. The name of the variable in the species characteristics that contains the species codes. Defaults to \code{"SpeciesCode"}.}

\item{generic_species_file}{Optional character string or data frame. Must be either a data frame or specify the filepath to generic species information (accepted filetypes are RDS, Rdata, and CSV). If this is \code{NULL}, it will inherit the value of \code{species_file}. Defaults to \code{NULL}.}

\item{source}{Character string (DEPRECATED). This argument is ignored but left here as a way to prevent legacy code from producing errors.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A data frame matching the format of height indicators in TerrADat.
}
\description{
This function calculates the full set of species-specific indicators that are standard for TerrADat.

This depends on the species characteristics used being those found in tblNationalPlants.

If there are indicators that cannot be calculated from the current data, those will be returned with a value of 0. For example, if the variable called Invasive does nto appear in the data or species characteristics provided, the variable NumSpp_Invasive will still be part of the output but all records will have a value of 0.
If the required variables for an indicator are present in the inputs, then every value for that indicator will be numeric, including using 0 where a PrimaryKey had no records with qualifying values for that indicator, e.g., if the variable Invasive is present at all but no records have a non-NA value in that variable, it will be assumed that all those species were determined to be non-invasive and therefore the returned value should be 0 rather than the NA associated with truly missing records.
These assumptions are less likely to hold true when applying this function to data from multiple sources simultaneously. For example, NumSpp_PreferredForb depends on per-state assignments and if the input data includes a state which has populated the SG_Group variable and one which has not, the most appropriate output value would be NA for the state without SG_Group assignments but the fact that the variable is present at all will produce 0s instead.
}
