% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.nri.R
\name{table_name}
\alias{table_name}
\title{Retrieve NRI table names from text}
\usage{
table_name(nri_path)
}
\arguments{
\item{nri_path}{Folder path. The path where the nri txt files are stored.}
}
\value{
Character string of the nri table names.
}
\description{
Retrieves the table names using the specified file referring to the folder containing NRI txt files
}
