% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tall_rangeland_health_sample}
\alias{tall_rangeland_health_sample}
\title{Tall rangeland health data}
\description{
Sample of tall rangeland health data produced by the gather_rangeland_health function.
}
\keyword{data}
\keyword{tall}
