% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tall_soil_stability_sample}
\alias{tall_soil_stability_sample}
\title{Tall soil stability data}
\description{
Sample of tall soil stability data produced by the gather_soil_stability function.
}
\keyword{data}
\keyword{tall}
