% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{tall_species}
\alias{tall_species}
\title{Create distinct species inventory records from semicolon-separated values}
\usage{
tall_species(species_inventory_detail)
}
\arguments{
\item{species_inventory_detail}{Data frame. The detail species inventory data to split into multiple records. Must contain the variables PrimaryKey, RecKey, and SpeciesList. SpeciesList must contain character strings composed of the species codes separated by semicolons, e.g., "ARTRW8;PRGL2;ARTR4".}
}
\value{
A data frame with the variables PrimaryKey, RecKey, and Species where each record represents one species code and its associated PrimaryKey and RecKey values.
}
\description{
Some species inventory data stored all species codes in a single character string as semicolon-separated values. This will create distinct records for each code.
This will only work with a data frame that has the variables PrimaryKey, RecKey, and SpeciesList.
}
