% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tall_species_inventory_sample}
\alias{tall_species_inventory_sample}
\title{Tall species inventory data}
\description{
Sample of tall species inventory data produced by the gather_species_inventory function.
}
\keyword{data}
\keyword{tall}
