% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terradactyl.R
\docType{package}
\name{terradactyl}
\alias{terradactyl-package}
\alias{terradactyl}
\title{terradactyl: A package for terrestrial data commons tools}
\description{
This package provides functionality for gathering terrestrial core methods
data, calculating standard variables and indicators, and producing
standard outputs for data repositories and models.
}
\references{
\href{https://doi.org/10.1016/j.ecolind.2022.109511}{McCord, S.E., J.R. Brehm, S.H. Burnett, C. Dietrich, B. Edwards, M. Hernandez-Narvaez, E. Kachergis, L.J. Metz, F. Pierson, K.S. Ramirez, N.G. Stauffer, N.P. Webb, C.E. Tweedie. A framework and toolset for standardizing ecosystem indicators. Ecological Indicators, vol 144, 2022.}
}
\author{
Sarah McCord \email{sarah.mccord@usda.gov}
}
