% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_specialty_functions.R
\name{lpi_calc_usgs}
\alias{lpi_calc_usgs}
\title{Line-point intercept indicator calculations}
\usage{
lpi_calc_usgs(header, lpi_tall, species_file, source,
  overwrite_generic_species = FALSE, by_year = TRUE,
  by_sampleperiod = 0)
}
\arguments{
\item{header}{Data frame. Use the data frame from the \code{header_build()}
output. Must be a .Rdata file.}

\item{lpi_tall}{Character. File path to a tall/long-format data frame. Use the data frame
from the \code{gather_lpi()} output. Must be a .Rdata file.}

\item{species_file}{Character. The full file path (including file extension)
to the file containing the species list}

\item{source}{Character. Specifies data source, "AIM", "LMF", "I&M", "DIMA"}

\item{overwrite_generic_species}{Logical. Updates attributes of generic species
codes to those contained in tblSpeciesGeneric.Defaults to \code{FALSE}.}

\item{by_year}{Logical. If \code{TRUE} then results will be reported further
grouped by year using the \code{FormDate} field from the data forms.
Defaults to \code{TRUE}.}

\item{by_sampleperiod}{Numeric >=\code{0} indicating the number of days in
the desired sampling window. If >\code{0}, results will be reported further
grouped by sampling period, where \code{FormDate} values for each plot that
fall within the given number of days of each other are considered a single
sampling period. If =\code{0}, no sample period grouping is performed.
Defaults to \code{0}.}
}
\value{
A \code{tbl} of indicators of either tall or wide format.
}
\description{
Calculate the percent cover by plot, and optionally line and year,
for ground cover, total foliar cover, and a suite of plant traits including
photosynthetic pathway, noxious status, duration, growth habit sub (e.g. forb,
grass, shrub), growth habit (woody and non-woody), and native status. Returns
values for all LPI hits and first LPI hits. Different from \code{lpi_calc()}
because it includes native species status and photosynthetic pathway in the
indicators calculated, and because it allows the user to input a separate
species file.
}
