% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulated_species.R
\name{accumulated_species}
\alias{accumulated_species}
\title{Accumulated species across all methods, by height and cover}
\usage{
accumulated_species(lpi_tall, height_tall, spp_inventory_tall, header,
  species_file = "", ...)
}
\arguments{
\item{lpi_tall}{Source of lpi Rdata file}

\item{height_tall}{Source of height Rdata file}

\item{header}{Source of header Rdata file}

\item{species_file}{File path to species file if you want species attributes or updated species. Geodatabase or csv allowed.}

\item{...}{Filtering expression to subset the number of plots}

\item{species_inventory_tall}{Source of species inventory Rdata file}
}
\description{
Accumulated species across all methods, by height and cover
}
