% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{adjust_species_attributes}
\alias{adjust_species_attributes}
\title{Harmonize species attributes with TerrADat needs}
\usage{
adjust_species_attributes(data, verbose = FALSE)
}
\description{
Harmonize species attributes with TerrADat needs
}
