% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AERO.R
\name{aero}
\alias{aero}
\title{Calculate AERO Inputs}
\usage{
aero(
  lpi_tall,
  gap_tall,
  height_tall,
  header,
  texture_raster = "",
  folder_location = "~/DataCommons/AERO/"
)
}
\arguments{
\item{lpi_tall}{Table. Line-point intercept data in tall format}

\item{gap_tall}{Table. Gap data in tall format}

\item{height_tall}{Table. Height data in tall format}

\item{header}{Table. Contains PrimaryKey, Latitude, and Longitude}

\item{texture_raster}{Raster. Soil texture layer with sand and clay percentages}

\item{folder_location}{Character. Location for function to save AERO input files}
}
\value{
AERO input files and an input_summary table which summarizes all input values in a single place.
}
\description{
Function for creating inputs to the AERO wind erosion model, requires height, line-point intercept, canopy gap, and soil texture observations for each plot.
}
