% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R, R/gathering.R
\name{build_terradat_indicators}
\alias{build_terradat_indicators}
\alias{build_lmf_indicators}
\alias{build_indicators}
\alias{lpi_calc}
\alias{gap_calc}
\alias{height_calc}
\alias{spp_inventory_calc}
\alias{soil_stability_calc}
\alias{gather_header_terradat}
\alias{gather_header_lmf}
\alias{gather_header_nri}
\alias{gather_header}
\title{Build AIM Indicators Tables and Feature Classes}
\usage{
build_terradat_indicators(
  header,
  dsn,
  species_file,
  lpi_tall = NULL,
  gap_tall = NULL,
  height_tall = NULL,
  spp_inventory_tall = NULL,
  soil_stability_tall = NULL,
  ...,
  verbose = FALSE
)

build_lmf_indicators(
  header,
  source,
  dsn,
  species_file,
  lpi_tall,
  gap_tall,
  height_tall,
  spp_inventory_tall,
  soil_stability_tall,
  ...,
  generic_species_file = NULL,
  verbose = FALSE
)

build_indicators(
  header,
  source,
  dsn = NULL,
  lpi_tall,
  species_file,
  gap_tall,
  height_tall,
  spp_inventory_tall,
  soil_stability_tall,
  ...,
  generic_species_file = NULL,
  verbose = FALSE
)

lpi_calc(
  header = NULL,
  lpi_tall = NULL,
  species_file,
  source,
  dsn = NULL,
  generic_species_file = NULL,
  digits = 1,
  verbose = FALSE
)

gap_calc(header, gap_tall, verbose = FALSE)

height_calc(
  header,
  height_tall,
  species_file = species_file,
  source,
  generic_species_file = NULL,
  verbose = FALSE
)

spp_inventory_calc(
  header,
  spp_inventory_tall,
  species_file,
  source,
  generic_species_file = NULL,
  verbose = FALSE
)

soil_stability_calc(soil_stability_tall, verbose = FALSE)

gather_header_terradat(
  dsn = NULL,
  tblPlots = NULL,
  date_tables = NULL,
  ...,
  verbose = FALSE
)

gather_header_lmf(dsn = NULL, ..., verbose = FALSE)

gather_header_nri(dsn = NULL, speciesstate, ..., verbose = FALSE)

gather_header(
  dsn = NULL,
  source,
  tblPlots = NULL,
  date_tables = NULL,
  speciesstate = NULL,
  ...,
  autoQC = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{header}{Dataframe. Plot header containing plot metadata}

\item{dsn}{String File path to the TerrADat database.}

\item{...}{Query in grepl format that subsets plots.}

\item{source}{String. Specifies data source, \code{"AIM", "LMF"}}
}
\value{
A \code{tbl} of indicators of either tall or wide format.
}
\description{
Build AIM Indicators Tables and Feature Classes

export gather_header_survey123
rdname aim_gdb
}
