% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R
\name{header_build_terradat}
\alias{header_build_terradat}
\alias{header_build_lmf}
\alias{header_build}
\alias{lpi_calc}
\alias{gap_calc}
\alias{height_calc}
\alias{spp_inventory_calc}
\alias{soil_stability_calc}
\alias{build_terradat_indicators}
\alias{build_lmf_indicators}
\title{Build AIM Indicators Tables and Feature Classes}
\usage{
header_build_terradat(dsn, ...)

header_build_lmf(dsn, ...)

header_build(dsn, source, ...)

lpi_calc(header, dsn, source)

gap_calc(header, dsn, source)

height_calc(header, dsn, source)

spp_inventory_calc(header, dsn, source)

soil_stability_calc(header, dsn, source)

build_terradat_indicators(dsn, source, ...)

build_lmf_indicators(dsn, source, ...)
}
\arguments{
\item{dsn}{String File path to the TerrADat database.}

\item{...}{Query in grepl format that subsets plots.}

\item{source}{String. Specifies data source, \code{"AIM", "LMF"}}

\item{header}{Dataframe. Plot header containing plot metadata}
}
\value{
A \code{tbl} of indicators of either tall or wide format.
}
\description{
Build AIM Indicators Tables and Feature Classes
}
