% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R, R/gathering.R
\name{lpi_calc}
\alias{lpi_calc}
\alias{gap_calc}
\alias{height_calc}
\alias{spp_inventory_calc}
\alias{soil_stability_calc}
\alias{build_terradat_indicators}
\alias{build_lmf_indicators}
\alias{build_indicators}
\alias{gather_header_terradat}
\alias{gather_header_lmf}
\alias{gather_header_nri}
\alias{gather_header}
\title{Build AIM Indicators Tables and Feature Classes}
\usage{
lpi_calc(header, lpi_tall, species_file, source, dsn, verbose = TRUE)

gap_calc(header, gap_tall, verbose = FALSE)

height_calc(header, height_tall, species_file = species_file, source)

spp_inventory_calc(header, spp_inventory_tall, species_file, source)

soil_stability_calc(header, soil_stability_tall)

build_terradat_indicators(
  header,
  source,
  dsn,
  species_file,
  lpi_tall,
  gap_tall,
  height_tall,
  spp_inventory_tall,
  soil_stability_tall,
  ...
)

build_lmf_indicators(
  header,
  source,
  dsn,
  species_file,
  lpi_tall,
  gap_tall,
  height_tall,
  spp_inventory_tall,
  soil_stability_tall,
  ...
)

build_indicators(
  header,
  source,
  dsn,
  lpi_tall,
  species_file,
  gap_tall,
  height_tall,
  spp_inventory_tall,
  soil_stability_tall,
  ...
)

gather_header_terradat(dsn = NULL, tblPlots = NULL, date_tables = NULL, ...)

gather_header_lmf(dsn = NULL, ...)

gather_header_nri(dsn = NULL, speciesstate, ...)

gather_header(
  dsn = NULL,
  source,
  tblPlots = NULL,
  date_tables = NULL,
  speciesstate = NULL,
  ...,
  autoQC = TRUE
)
}
\arguments{
\item{header}{Dataframe. Plot header containing plot metadata}

\item{source}{String. Specifies data source, \code{"AIM", "LMF"}}

\item{dsn}{String File path to the TerrADat database.}

\item{...}{Query in grepl format that subsets plots.}
}
\value{
A \code{tbl} of indicators of either tall or wide format.
}
\description{
Build AIM Indicators Tables and Feature Classes

export gather_header_survey123
rdname aim_gdb
}
