% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R
\name{header_build_terradat}
\alias{header_build_terradat}
\alias{header_build_lmf}
\alias{header_build}
\alias{lpi_calc}
\alias{gap_calc}
\alias{height_calc}
\alias{spp_inventory_calc}
\alias{soil_stability_calc}
\alias{build_terradat_indicators}
\alias{build_lmf_indicators}
\alias{build_indicators}
\title{Build AIM Indicators Tables and Feature Classes}
\usage{
header_build_terradat(dsn, ...)

header_build_lmf(dsn, ...)

header_build(dsn, source, ...)

lpi_calc(header, lpi_tall, species_file, source)

gap_calc(header, gap_tall)

height_calc(header, height_tall, species_file = species_file, source)

spp_inventory_calc(header, spp_inventory_tall, species_file)

soil_stability_calc(header, soil_stability_tall)

build_terradat_indicators(dsn, source, species_file, lpi_tall, gap_tall,
  height_tall, spp_inventory_tall, soil_stability_tall, ...)

build_lmf_indicators(dsn, source, species_file, lpi_tall, gap_tall,
  height_tall, spp_inventory_tall, soil_stability_tall, ...)

build_indicators(dsn, source, lpi_tall, species_file, gap_tall,
  height_tall, spp_inventory_tall, soil_stability_tall, ...)
}
\arguments{
\item{dsn}{String File path to the TerrADat database.}

\item{...}{Query in grepl format that subsets plots.}

\item{source}{String. Specifies data source, \code{"AIM", "LMF"}}

\item{header}{Dataframe. Plot header containing plot metadata}
}
\value{
A \code{tbl} of indicators of either tall or wide format.
}
\description{
Build AIM Indicators Tables and Feature Classes
}
