% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R
\name{header_build}
\alias{header_build}
\alias{lpi_calc}
\alias{gap_calc}
\alias{height_calc}
\alias{spp_inventory_calc}
\alias{soil_stability_calc}
\alias{build.terradat.indicators}
\title{Build AIM Indicators Tables and Feature Classes}
\usage{
header_build(dsn, ...)

lpi_calc(header, dsn, layer)

gap_calc(header, dsn, layer)

height_calc(header, dsn)

spp_inventory_calc(header, dsn)

soil_stability_calc(header, dsn)

build.terradat.indicators(dsn, layer, ...)
}
\arguments{
\item{dsn}{String File path to the TerrADat database.}
}
\value{
A \code{tbl} of indicators of either tall or wide format.
}
\description{
Build AIM Indicators Tables and Feature Classes
}
