% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.nri.R
\name{assign_pkey_nri}
\alias{assign_pkey_nri}
\title{Assign PrimaryKeys to nri data and remove sensitive information}
\usage{
assign_pkey_nri(df)
}
\arguments{
\item{df}{list of nri dataframes with header and df names assigned}
}
\value{
nri df list with PrimaryKey
}
\description{
Assign PrimaryKeys to nri data and remove sensitive information
}
