% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_cover_indicators.R
\name{core_cover_indicators}
\alias{core_cover_indicators}
\title{Core Cover Indicators
Generates cover calculations from the Line-point intercept method that are likely generalizable across programs (e.g., TotalFoliarCover, TotalLitterCover, BareSoilCover, cover by GrowthHabit, etc)}
\usage{
core_cover_indicators(lpi_species)
}
\arguments{
\item{lpi_species}{Dataframe in environment. An LPI tall table generated from \code{gather_lpi()} and \code{species_join()}.}
}
\description{
Core Cover Indicators
Generates cover calculations from the Line-point intercept method that are likely generalizable across programs (e.g., TotalFoliarCover, TotalLitterCover, BareSoilCover, cover by GrowthHabit, etc)
}
\examples{
library(terradactyl)
cover <- core_cover_indicators(lpi_species)
}
