% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.indicators.R
\name{cover_indicators}
\alias{cover_indicators}
\alias{pct_cover_between_plant}
\alias{pct_cover_all_ground}
\alias{pct_cover_total_foliar}
\alias{pct_cover_bare_soil}
\alias{pct_cover_litter}
\alias{pct_cover_live}
\alias{pct_cover_species}
\title{Percent Cover Indicators}
\usage{
pct_cover_between_plant(
  lpi_tall,
  tall = FALSE,
  by_year = FALSE,
  by_line = FALSE
)

pct_cover_all_ground(lpi_tall, tall = FALSE, by_year = FALSE, by_line = FALSE)

pct_cover_total_foliar(
  lpi_tall,
  tall = FALSE,
  by_year = FALSE,
  by_line = FALSE
)

pct_cover_bare_soil(lpi_tall, tall = FALSE, by_year = FALSE, by_line = FALSE)

pct_cover_litter(lpi_tall, tall = FALSE, by_year = FALSE, by_line = FALSE)

pct_cover_live(
  lpi_tall,
  tall = FALSE,
  by_year = FALSE,
  by_line = FALSE,
  hit = "any",
  ...
)

pct_cover_species(
  lpi_tall,
  tall = TRUE,
  by_year = FALSE,
  by_line = FALSE,
  hit = "any"
)
}
\arguments{
\item{lpi_tall}{A tall/long-format data frame. Use the data frame \code{"layers"} from the \code{gather.lpi()} output.}

\item{tall}{Logical. If \code{TRUE} then output will be in tall format}

\item{by_year}{Logical. If \code{TRUE} then results will be reported further grouped by year using the \code{DateModified} field from the data forms. Defaults to \code{FALSE}.}

\item{by_line}{Logical. If \code{TRUE} then results will be reported further grouped by line using the \code{LineID} and \code{LineKey} fields from the data forms. Defaults to \code{FALSE}.}

\item{hit}{String. If \code{"first"} then only top LPI hits are included. If \code{"any"} then any hit values are included.}

\item{...}{Optional bare variable names. Names of variables to include as part of grouping e.g. \code{GrowthHabitSub} to calculate percent cover by growth habits or \code{GrowthHabitSub, Duration} to calculate percent cover for categories like perennial forbs, annual graminoids, etc.}
}
\value{
A \code{tbl} of either wide or tall format.
}
\description{
Calculate the percent cover  indicators by plot or line for variables or combinations of variables.This is a family of standard indicator variables to examine total foliar cover, bare soil, litter cover, and other ground cover indicators. To compute cover by species, growth habit and duration, or other custom line-point intercept combinations, see \code{pct_cover()}.
}
