% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.height.R
\name{gather.height}
\alias{gather.height}
\title{Convert wide-format TerrADat height data to a tall, tidy format}
\usage{
gather.height(dsn, species.characteristics = TRUE, species.file = "",
  species.code = "SpeciesCode",
  species.growth.habit.code = "GrowthHabitSub",
  species.duration = "Duration", growth.habit.file = "",
  growth.habit.code = "Code")
}
\description{
Convert wide-format TerrADat height data to a tall, tidy format
}
