% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.height.R
\name{gather.height.terradat}
\alias{gather.height.terradat}
\alias{gather.height.lmf}
\alias{gather.height}
\title{Convert wide-format TerrADat height data to a tall, tidy format}
\usage{
gather.height.terradat(dsn)

gather.height.lmf(dsn, file.type = "gdb")

gather.height(dsn, file.type = "gdb", source)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extension) of the geodatabase containing the table of interest.}

\item{source}{Character string. Identify the source data ("TerrADat", "LMF", "NRI") for gathering}
}
\value{
A data frames containing the data from the height measurements.
}
\description{
Given a list of data frames containing tblSites, tblPlots, tblLines, tblLPIHeader, and tblLPIDetail, create a tall format data frame for canopy data from LPI and one for heights from the specialized height fields.
}
