% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_all.R
\name{gather_all}
\alias{gather_all}
\title{Build tall tables for all AIM methods}
\usage{
gather_all(dsn, folder)
}
\arguments{
\item{dsn}{String. Filepath to data}

\item{folder}{Folder location for tall Rdata files}
}
\description{
Build tall tables for all AIM methods
}
