% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_gap}
\alias{gather_gap}
\title{export gather_gap_survey123
rdname gather_gap
Convert wide gap data into a long/tall format.}
\usage{
gather_gap(
  dsn = NULL,
  file_type = "gdb",
  source,
  tblGapHeader = NULL,
  tblGapDetail = NULL,
  POINT = NULL,
  GINTERCEPT = NULL,
  autoQC = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{source}{Character string. The data source format. Must be one of \code{"AIM"}, \code{"TerrADat"}, \code{"DIMA"}, or \code{"LMF"} (case independent).}
}
\description{
Given wide format line-point intercept or height data, create a tall
format data frame usable by other terradactyl functions. For additional information about
arguments, see the documentation for the functions \code{\link[terradactyl:gather_gap_terradat]{gather_gap_terradat()}} and \code{\link[terradactyl:gather_gap_lmf]{gather_gap_lmf()}}.
}
\details{
The \code{source} argument determines which other arguments are used or ignored.

When \code{source} is one of \code{"AIM"}, \code{"TERRADAT"}, or \code{"DIMA"}
then the arguments \code{tblGapHeader} and \code{tblGapDetail} are both considered.

When \code{source} is \code{"LMF"} then the arguments \code{GINTERCEPT} and \code{POINT} are considered.

Regardless of the value of \code{source}, the data sources represented by those other arguments are required. The simplest way to provide them is to provide the filepath to a geodatabase as \code{dsn} with each of those feature classes appearing by the same name as the corresponding argument in that geodatabase.
}
