% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_gap}
\alias{gather_gap}
\title{export gather_gap_survey123
rdname gather_gap}
\usage{
gather_gap(
  dsn = NULL,
  file_type = "gdb",
  source,
  tblGapHeader = NULL,
  tblGapDetail = NULL,
  POINT = NULL,
  GINTERCEPT = NULL,
  autoQC = TRUE,
  verbose = FALSE
)
}
\description{
export gather_gap_survey123
rdname gather_gap
}
