% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_gap.R
\name{gather.gap.terradat}
\alias{gather.gap.terradat}
\alias{gather.gap.lmf}
\alias{gather.gap}
\title{Gather TerrADat Gap data into tall/long data frames}
\usage{
gather.gap.terradat(dsn)

gather.gap.lmf(dsn, file.type = "gdb")

gather.gap(dsn, file.type = "gdb", source)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extension) of the geodatabase containing the table of interest.}

\item{file.type}{Character string. Type of file, text or geodatabase, to read from.}

\item{source}{Character string. The original source of the data. "TerrAdat", "AIM", "DIMA", "LMF", "NRI" are all valide options.}
}
\value{
A data frames containing the data from the Gap intercepts data in tall format.
}
\description{
Given a list of data frames containing tblSites, tblPlots, tblLines, tblLPIHeader, and tblLPIDetail, create a tall format data frame for Gap.
}
