% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_gap.R
\name{gather_gap_terradat}
\alias{gather_gap_terradat}
\alias{gather_gap_lmf}
\alias{gather_gap}
\title{Gather TerrADat Gap data into tall/long data frame}
\usage{
gather_gap_terradat(dsn)

gather_gap_lmf(dsn, file_type = "gdb")

gather_gap(dsn, file_type = "gdb", source)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extension) of the geodatabase containing the table of interest.}

\item{file_type}{Character string. Type of file, text or geodatabase, to read from.}

\item{source}{Character string. The original source of the data. "TerrAdat", "AIM", "DIMA", "LMF", "NRI" are all valide options.}
}
\value{
A data frames containing the data from the Gap intercepts data in tall format.
}
\description{
Given a list of data frames containing tblSites, tblPlots, tblLines, tblLPIHeader, and tblLPIDetail, create a tall format data frame for Gap.
}
