% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_gap.R
\name{gather_gap}
\alias{gather_gap}
\alias{gather_gap_terradat}
\alias{gather_gap_lmf}
\title{Convert gap data into a tall, tidy data frame}
\usage{
gather_gap_terradat(dsn = NULL, tblGapDetail = NULL, tblGapHeader = NULL)

gather_gap_lmf(dsn = NULL, file_type = NULL, GINTERCEPT = NULL, POINT = NULL)

gather_gap(
  dsn = NULL,
  file_type = "gdb",
  source,
  tblGapHeader = NULL,
  tblGapDetail = NULL,
  POINT = NULL,
  GINTERCEPT = NULL,
  Gap_0 = NULL,
  GapDetail_1 = NULL
)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file
extension) of the geodatabase or text file containing the table of interest.
This field is unnecessary if you provide either both of tblGapDetail and
tblGapHeader (AIM/DIMA/TerrADat) or both of GINTERCEPT and POINT (LMF/NRI).}

\item{tblGapDetail}{Dataframe of the data structure tblGapDetail from the
DIMA database with the addition of PrimaryKey and DBKey fields. Use with
tblGapHeader when data source is AIM, DIMA, or TerrADat; alternately provide
dsn.}

\item{tblGapHeader}{Dataframe of the data structure tblGapHeader from the
DIMA database with the addition of PrimaryKey and DBKey fields. Use with
tblGapDetail when data source is AIM, DIMA, or TerrADat; alternately provide
dsn.}

\item{file_type}{Character string that denotes the source file type of the
LMF/NRI data, \code{"gdb"} or \code{"txt"}. Not necessary for
AIM/DIMA/TerrADat, or if GINTERCEPT and POINT are provided.}

\item{GINTERCEPT}{Dataframe of the data structure GINTERCEPT from the LMF/NRI
database. Use with POINT when data source is LMF or NRI; alternately provide
dsn.}

\item{POINT}{Dataframe of the data structure POINT from the LMF database.
Use with GINTERCEPT when data source if LMF or NRI; alternately provide dsn.}

\item{source}{Character string. The data source format,
\code{"AIM", "TerrADat", "DIMA", "LMF", "NRI"} (case independent).}
}
\value{
A tall data frame containing the data from the gap measurements.
}
\description{
Given wide format gap data, create a tall format data frame
usable by other terradactyl functions.
}
\examples{
gather_gap(dsn = "Path/To/AIM_Geodatabase.gdb",
           source = "AIM")
gather_gap(dsn = "Path/To/LMF_Geodatabase.gdb",
           source = "LMF")

aim_gapdetail <- read.csv("Path/To/tblGapDetail.csv")
aim_gapheader <- read.csv("Path/To/tblGapHeader.csv")
gather_gap(source = "AIM",
           tblGapDetail = aim_gapdetail,
           tblGapHeader = aim_gapheader)

lmf_gintercept <- read.csv("Path/To/GINTERCEPT.csv")
lmf_point <- read.csv("Path/To/POINT.csv")
gather_gap(source = "LMF",
           GINTERCEPT = lmf_gintercept,
           POINT = lmf_point)
}
\seealso{
Other <gather>: 
\code{\link{gather_all}()},
\code{\link{gather_height}()},
\code{\link{gather_lpi}()},
\code{\link{gather_plot_characterization}()},
\code{\link{gather_rangeland_health}()},
\code{\link{gather_soil_horizon}()},
\code{\link{gather_soil_stability}()},
\code{\link{gather_soil_summary}()},
\code{\link{gather_species_inventory}()}
}
\concept{<gather>}
