% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_header}
\alias{gather_header}
\title{export gather_header_survey123
rdname aim_gdb
Gather plot-level header data}
\usage{
gather_header(
  dsn = NULL,
  source,
  tblPlots = NULL,
  date_tables = NULL,
  speciesstate = NULL,
  ...,
  autoQC = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. This must be the filepath
to a geodatabase which contains the relevant feature classes, e.g.
\code{"C:/DATA/AIM.GDB"}. This is optional only when \code{source} is
\code{"AIM"}, \code{"TERRADAT"}, or \code{"DIMA"}. If \code{source} is
\code{"LMF"} or \code{"NRI"} this is required. Defaults to \code{NULL}
which results in it being ignored by the function.}

\item{source}{Character string. This specifies the expected data format(s)
and determines which specialized gather function will be used. It must be
one of \code{"AIM"}, \code{"TERRADAT"}, \code{"LMF"}, \code{"DIMA"} or
\code{"NRI"}.}

\item{tblPlots}{Optional data frame only used if \code{source} is
\code{"AIM"}, \code{"TERRADAT"}, or \code{"DIMA"}. If provided, this must
contain the expected plot metadata. If \code{NULL} then the argument
\code{dsn} must be provided. Defaults to \code{NULL}.}

\item{date_tables}{Optional (contingent on other arguments) character vector
or list of data frames. Only used if \code{source} is \code{"AIM"},
\code{"TERRADAT"}, or \code{"DIMA"}. This specifies the tables to extract
date information from. If \code{dsn} is not \code{NULL} AND the geodatabase
contains at least one feature class from the set tblLPIHeader,
tblGapHeader, "tblSpecRichHeader", this argument is optional and can be
left as \code{NULL}. Otherwise, if \code{dsn} is not \code{NULL} and the
desired tables to use are named anything else, this must be a vector of
character strings specifying the names of the relevant feature classes. If
\code{tblPlots} is being used, then this must be a list of data frames
containing the relevant data. Defaults to \code{NULL}.}

\item{speciesstate}{Optional. Used by NRI.}

\item{...}{Additional optional filtering expressions to be used with
\code{dplyr::filter()} to restrict the processing and output to only a
subset of input data.}

\item{autoQC}{Logical (currently disabled). If \code{TRUE} then automatic
quality control functions will be applied to the data before returning the
output. Defaults to \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A long-format data frame of header data.
}
\description{
This is a wrapper function for specialized functions which read
in metadata from AIM sampling used as headers for various methods and returns
it as a long-format data frame suitable for use in indicator calculations
with the package \code{terradactyl}. Supported formats include those used in
AIM (Assessment, Inventory, and Monitoring) and TerrADat (Terrestrial AIM
Database), LMF (Landscape Monitoring Framework), and DIMA (the Database for
Inventory, Monitoring, and Assessment). For additional information about
arguments, see the documentation for the functions
\code{gather_header_terradat()}, \code{gather_header_lmf()}, and
\code{gather_header_nri()}
}
