% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_header_lmf}
\alias{gather_header_lmf}
\title{Gather LMF plot-level header data}
\usage{
gather_header_lmf(
  dsn = NULL,
  POINT = NULL,
  POINTCOORDINATES = NULL,
  GPS = NULL,
  ESFSG = NULL,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{POINT}{Data frame or character string. If a data frame, must contain the variables PrimaryKey, SpeciesState, COUNTY, and STATE. If a character string, must either correspond to the filepath to a CSV or RDATA file containing a table with those variables or the name of the feature class in the geodatabase provided as \code{dsn} with those variables. If \code{NULL}, the function will attempt to find a feature class called POINT (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{POINTCOORDINATES}{Data frame or character string. If a data frame, must contain the variables PrimaryKey, LocationType, REPORT_LATITUDE, and REPORT_LONGITUDE. If a character string, must either correspond to the filepath to a CSV or RDATA file containing a table with those variables or the name of the feature class in the geodatabase provided as \code{dsn} with those variables. If \code{NULL}, the function will attempt to find a feature class called POINTCOORDINATES (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{GPS}{Data frame or character string. If a data frame, must contain the variables PrimaryKey, CAPDATE, and ELEVATION. If a character string, must either correspond to the filepath to a CSV or RDATA file containing a table with those variables or the name of the feature class in the geodatabase provided as \code{dsn} with those variables. If \code{NULL}, the function will attempt to find a feature class called GPS (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{ESFSG}{Data frame or character string. If a data frame, must contain the variables COVERAGE, END_MARK, START_MARK, ESFSG_MLRA, ESFSG_SITE, and ESFSG_STATE (the variable ESFSG_PREFIX is optional). If a character string, must either correspond to the filepath to a CSV or RDATA file containing a table with those variables or the name of the feature class in the geodatabase provided as \code{dsn} with those variables. If \code{NULL}, the function will attempt to find a feature class called ESFSG (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{...}{Additional optional filtering expressions passed to
\code{\link[dplyr:filter]{dplyr::filter()}} to restrict the processing and output to only a
subset of input data.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A long-format data frame of header data.
}
\description{
This reads in metadata from AIM sampling used as headers for
various methods and returns it as a long-format data frame suitable for use
in indicator calculations with the package \code{terradactyl}. The required
feature classes in the geodatabase specified with \code{dsn} are POINT,
POINTCOORDINATES, GPS, and ESFSG. The expected formats for
the input data are those used in the Landscape Monitoring Framework.
}
\examples{
# Basic use assuming that all expected feature classes appear in lmf_data.gdb
gather_header_lmf(dsn = "data_path/lmf_data.gdb")

# Using a filtering expression to restrict the processing and output to only
# records where the SpeciesState variable in POINT contained either of the
# values "CO" and "NM".
gather_header_lmf(dsn = "data_path/lmf_data.gdb",
                  SpeciesState \%in\% c("CO", "NM"))

}
