% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_height.R
\name{gather_height}
\alias{gather_height}
\alias{gather_height_terradat}
\alias{gather_height_lmf}
\alias{gather_height_survey123}
\title{Convert height data into a tall, tidy data frame}
\usage{
gather_height_terradat(dsn = NULL, tblLPIDetail = NULL, tblLPIHeader = NULL)

gather_height_lmf(dsn = NULL, file_type = "gdb", PASTUREHEIGHTS = NULL)

gather_height_survey123(LPI_0, LPIDetail_1)

gather_height(
  dsn = NULL,
  file_type = "gdb",
  source,
  tblLPIDetail = NULL,
  tblLPIHeader = NULL,
  PASTUREHEIGHTS = NULL,
  LPI_0 = NULL,
  LPIDetail_1 = NULL
)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file
extension) of the geodatabase or text file containing the table of interest.
This field is unnecessary if you provide either both of tblLPIDetail and
tblLPIHeader (AIM/DIMA/TerrADat) or PASTUREHEIGHTS (LMF/NRI).}

\item{tblLPIDetail}{Dataframe of the data structure tblLPIDetail from the
DIMA database with the addition of PrimaryKey and DBKey fields. Use with
tblLPIHeader when data source is AIM, DIMA, or TerrADat; alternately provide
dsn.}

\item{tblLPIHeader}{Dataframe of the data structure tblLPIHeader from the
DIMA database with the addition of PrimaryKey and DBKey fields. Use with
tblLPIDetail when data source is AIM, DIMA, or TerrADat; alternately provide
dsn.}

\item{file_type}{Character string that denotes the source file type of the
LMF/NRI data, \code{"gdb"} or \code{"txt"}. Not necessary for
AIM/DIMA/TerrADat, or if PASTUREHEIGHTS is provided.}

\item{PASTUREHEIGHTS}{Dataframe of the data structure PASTUREHEIGHTS from the
LMF/NRI database; alternately provide dsn.}

\item{source}{Character string. The data source format,
\code{"AIM", "TerrADat", "DIMA", "LMF", "NRI"} (case independent).}
}
\value{
A tall data frame containing the data from the height measurements.
}
\description{
Given wide format line-point intercept data, create a tall
format data frame usable by other terradactyl functions.
}
\examples{
gather_height(dsn = "Path/To/AIM_Geodatabase.gdb",
              source = "AIM")
gather_height(dsn = "Path/To/LMF_Geodatabase.gdb",
              source = "LMF")

aim_lpidetail <- read.csv("Path/To/tblLPIDetail.csv")
aim_lpiheader <- read.csv("Path/To/tblLPIHeader.csv")
gather_height(source = "AIM",
              tblLPIDetail = aim_lpidetail,
              tblLPIHeader = aim_lpiheader)

lmf_heights <- read.csv("Path/To/PASTUREHEIGHTS.csv")
gather_height(source = "LMF",
              PASTUREHEIGHTS = lmf_heights)
}
\seealso{
Other <gather>: 
\code{\link{gather_all}()},
\code{\link{gather_gap}()},
\code{\link{gather_lpi}()},
\code{\link{gather_plot_characterization}()},
\code{\link{gather_rangeland_health}()},
\code{\link{gather_soil_horizon}()},
\code{\link{gather_soil_stability}()},
\code{\link{gather_soil_summary}()},
\code{\link{gather_species_inventory}()}
}
\concept{<gather>}
