% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_height}
\alias{gather_height}
\alias{gather_height_terradat}
\alias{gather_height_lmf}
\title{Convert height data into a tall, tidy data frame}
\usage{
gather_height_terradat(
  dsn = NULL,
  tblLPIDetail = NULL,
  tblLPIHeader = NULL,
  auto_qc_warnings = TRUE,
  verbose = FALSE
)

gather_height_lmf(
  dsn = NULL,
  file_type = "gdb",
  PASTUREHEIGHTS = NULL,
  verbose = FALSE
)

gather_height(
  dsn = NULL,
  file_type = "gdb",
  source,
  tblLPIDetail = NULL,
  tblLPIHeader = NULL,
  PASTUREHEIGHTS = NULL,
  autoQC = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file
extension) of the geodatabase or text file containing the table of interest.
This field is unnecessary if you provide either both of tblLPIDetail and
tblLPIHeader (AIM/DIMA/TerrADat) or PASTUREHEIGHTS (LMF/NRI).}

\item{tblLPIDetail}{Dataframe of the data structure tblLPIDetail from the
DIMA database with the addition of PrimaryKey and DBKey fields. Use with
tblLPIHeader when data source is AIM, DIMA, or TerrADat; alternately provide
dsn.}

\item{tblLPIHeader}{Dataframe of the data structure tblLPIHeader from the
DIMA database with the addition of PrimaryKey and DBKey fields. Use with
tblLPIDetail when data source is AIM, DIMA, or TerrADat; alternately provide
dsn.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}

\item{file_type}{Character string that denotes the source file type of the
LMF/NRI data, \code{"gdb"} or \code{"txt"}. Not necessary for
AIM/DIMA/TerrADat, or if \code{PASTUREHEIGHTS} is provided.}

\item{PASTUREHEIGHTS}{Dataframe of the data structure PASTUREHEIGHTS from the
LMF/NRI database; alternately provide \code{dsn}.}

\item{source}{Character string. The data source format,
\code{"AIM", "TerrADat", "DIMA", "LMF", "NRI"} (case independent).}

\item{auto_qc}{Logical. If \code{TRUE} then AIM/DIMA/TerrADat data will be
checked for non-unique and orphaned records before doing processing. If any
are found, a warning will be triggered but the gather will still be carried
out. It is strongly recommended that any identified issues be addressed to
avoid incorrect records in the output. Defaults to \code{TRUE}.}
}
\value{
A tall data frame containing the data from the height measurements.
}
\description{
Given wide format line-point intercept data, create a tall
format data frame usable by other terradactyl functions.
}
\examples{
gather_height(dsn = "Path/To/AIM_Geodatabase.gdb",
              source = "AIM")
gather_height(dsn = "Path/To/LMF_Geodatabase.gdb",
              source = "LMF")

aim_lpidetail <- read.csv("Path/To/tblLPIDetail.csv")
aim_lpiheader <- read.csv("Path/To/tblLPIHeader.csv")
gather_height(source = "AIM",
              tblLPIDetail = aim_lpidetail,
              tblLPIHeader = aim_lpiheader)

lmf_heights <- read.csv("Path/To/PASTUREHEIGHTS.csv")
gather_height(source = "LMF",
              PASTUREHEIGHTS = lmf_heights)
}
\seealso{
Other <gather>: 
\code{\link{gather_gap}()},
\code{\link{gather_lpi}()},
\code{\link{gather_rangeland_health}()},
\code{\link{gather_soil_stability}()},
\code{\link{gather_species_inventory}()}
}
\concept{<gather>}
