% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_lpi}
\alias{gather_lpi}
\title{export gather_lpi_survey123
rdname gather_lpi
Function to convert LPI data into a long format.}
\usage{
gather_lpi(
  dsn = NULL,
  file_type = "gdb",
  source,
  tblLPIDetail = NULL,
  tblLPIHeader = NULL,
  PINTERCEPT = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. If provided, this must be the filepath
to a geodatabase which contains the feature classes tblLPIHeader and
tblLPIDetail or the feature class PINTERCEPT, e.g. \code{"C:/DATA/AIM.GDB"}. If this is \code{NULL}, then
the arguments \code{tblLPIHeader} and \code{tblLPIDetail} or the argument \code{PINTERCEPT} must be provided.
Defaults to \code{NULL}.}

\item{file_type}{Optional character string. This must be \code{"gdb"},
\code{"txt"}, or \code{"csv"} and must match the file type of \code{dsn}.
If \code{dsn} is \code{NULL} this argument is ignored.  Only used if \code{source} is \code{"LMF}. Defaults to
\code{"gdb"}.}

\item{source}{Character string. This specifies the expected data format(s)
and determines which specialized gather function will be used. It must be
one of \code{"AIM"}, \code{"TERRADAT"}, \code{"LMF"}, \code{"DIMA"} or
\code{"NRI"}. This is case-insensitive.}

\item{tblLPIDetail}{Optional data frame. If provided, this must contain the
expected LPI data in the format of the tblLPIDetail table used by AIM.  Only used if \code{source} is one of \code{"AIM"}, \code{"TERRADAT"}, or \code{"DIMA"}. If \code{NULL} then the argument
\code{dsn} must be provided. Defaults to \code{NULL}.}

\item{tblLPIHeader}{Optional data frame. If provided, this must contain the
expected metadata for the LPI data in the format of the tblLPIHeader table used by AIM. Only used if \code{source} is one of \code{"AIM"}, \code{"TERRADAT"}, or \code{"DIMA"}. If \code{NULL} then the argument \code{dsn} must be
provided. Defaults to \code{NULL}.}

\item{PINTERCEPT}{Optional (contingent) data frame. This is the data frame
containing the LPI data in the format of the PINTERCEPT table used by LMF.  Only used if \code{source} is \code{"LMF}.
If provided, this will be used even if \code{dsn} is also non-NULL.
Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
This is a wrapper function for specialized functions which read
in Line-Point Intercept data from AIM anbd LMF sampling and return
them as a long-format data frame suitable for use in indicator calculations
with the package \code{terradactyl}. Supported formats include those used in
AIM (Assessment, Inventory, and Monitoring) and TerrADat (Terrestrial AIM
Database), LMF (Landscape Monitoring Framework), and DIMA (the Database for
Inventory, Monitoring, and Assessment). For additional information about
arguments, see the documentation for the functions
\code{gather_lpi_terradat()} and \code{gather_lpi_lmf()}.
}
\examples{
# LPI data from a geodatabase in the format of the Terrestrial AIM Database
gather_lpi(dsn = "data_path/aim_data.gdb,
           source = "terradat")

# LPI data from a geodatabase in the format of the Landscape Monitoring Framework
gather_lpi(dsn = "data_path/lmf_data.gdb,
           source = "lmf")


}
