% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_lpi.R
\name{gather_lpi}
\alias{gather_lpi}
\alias{gather.lpi.terradat}
\alias{gather.lpi.lmf}
\alias{gather.lpi}
\title{Gather TerrADat LPI data into tall/long data frames}
\usage{
gather.lpi.terradat(dsn)

gather.lpi.lmf(dsn, file.type = "gdb")

gather.lpi(dsn, file.type = "gdb", source)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extension) of the geodatabase containing the table of interest.}

\item{source}{Character string. The data source format. \code("AIM", "NRI", "LMF, "TerrADat") are valid options.}
}
\value{
A data frames containing the data from the LPI pin intercepts
}
\description{
Given a list of data frames containing tblSites, tblPlots, tblLines, tblLPIHeader, and tblLPIDetail, create a tall format data frame for canopy data from LPI and one for heights from the specialized height fields.
}
\concept{<gather>}
