% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_lpi}
\alias{gather_lpi}
\alias{gather_lpi_terradat}
\alias{gather_lpi_lmf}
\title{Convert line-point intercept (LPI) data into a tall, tidy data frame}
\usage{
gather_lpi_terradat(
  dsn = NULL,
  tblLPIDetail = NULL,
  tblLPIHeader = NULL,
  auto_qc_warnings = TRUE,
  null_is_na = TRUE,
  verbose = FALSE
)

gather_lpi_lmf(dsn = NULL, file_type = "gdb", PINTERCEPT = NULL)

gather_lpi(
  dsn = NULL,
  file_type = "gdb",
  source,
  tblLPIDetail = NULL,
  tblLPIHeader = NULL,
  PINTERCEPT = NULL,
  autoQC = TRUE
)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file
extension) of the geodatabase or text file containing the table of interest.
This field is unnecessary if you provide either both of tblLPIDetail and
tblLPIHeader (AIM/DIMA/TerrADat) or PINTERCEPT (LMF/NRI).}

\item{tblLPIDetail}{Dataframe of the data structure tblLPIDetail from the
DIMA database with the addition of PrimaryKey and DBKey fields. Use with
tblLPIHeader when data source is AIM, DIMA, or TerrADat; alternately provide
dsn.}

\item{tblLPIHeader}{Dataframe of the data structure tblLPIHeader from the
DIMA database with the addition of PrimaryKey and DBKey fields. Use with
tblLPIDetail when data source is AIM, DIMA, or TerrADat; alternately provide
dsn.}

\item{null_is_na}{Logical. If \code{TRUE} then any instance of \code{"NULL"}
as a value for a layer in \code{tblLPIDetail} will be considered to be \code{NA}.
Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}

\item{file_type}{Character string that denotes the source file type of the
LMF/NRI data, \code{"gdb"} or \code{"txt"}. Not necessary for
AIM/DIMA/TerrADat, or if PINTERCEPT is provided.}

\item{PINTERCEPT}{Dataframe of the data structure PINTERCEPT from the LMF/NRI
database with the addition of PrimaryKey and DBKey fields. Use when source
is LMF or NRI; alternately provide dsn.}

\item{source}{Character string. The data source format,
\code{"AIM", "TerrADat", "DIMA", "LMF", "NRI"} (case independent).}

\item{auto_qc}{Logical. If \code{TRUE} then AIM/DIMA/TerrADat data will be
checked for non-unique and orphaned records before doing processing. If any
are found, a warning will be triggered but the gather will still be carried
out. It is strongly recommended that any identified issues be addressed to
avoid incorrect records in the output. Defaults to \code{TRUE}.}
}
\value{
A tall data frame containing the data from the LPI pin intercepts
}
\description{
Given wide format line-point intercept data, create a tall
format data frame usable by other terradactyl functions.
}
\examples{
gather_lpi(dsn = "Path/To/AIM_Geodatabase.gdb",
           source = "AIM")
gather_lpi(dsn = "Path/To/LMF_Geodatabase.gdb",
           source = "LMF")

aim_lpidetail <- read.csv("Path/To/tblLPIDetail.csv")
aim_lpiheader <- read.csv("Path/To/tblLPIHeader.csv")
gather_lpi(source = "AIM",
           tblLPIDetail = aim_lpidetail,
           tblLPIHeader = aim_lpiheader)

lmf_pintercept <- read.csv("Path/To/PINTERCEPT.csv")
gather_lpi(source = "LMF",
           RANGEHEALTH = lmf_pintercept)
}
\seealso{
Other <gather>: 
\code{\link{gather_all}()},
\code{\link{gather_gap}()},
\code{\link{gather_height}()},
\code{\link{gather_plot_characterization}()},
\code{\link{gather_rangeland_health}()},
\code{\link{gather_soil_horizon}()},
\code{\link{gather_soil_stability}()},
\code{\link{gather_species_inventory}()}
}
\concept{<gather>}
