% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_lpi_lmf}
\alias{gather_lpi_lmf}
\title{Function to convert LMF-format LPI data into a long format.}
\usage{
gather_lpi_lmf(
  dsn = NULL,
  PINTERCEPT = NULL,
  file_type = "gdb",
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{PINTERCEPT}{Optional data frame or character string. If provided, this must contain the
expected LPI data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called PINTERCEPT (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{file_type}{Deprecated. This argument is no longer functional or necessary and is kept for backwards compatibility with legacy code.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back diagnostic
information as console messages while it works. Defaults to \code{FALSE}.}
}
\description{
This reads in Line-Point Intercept from LMF sampling and returns it as a long-format data frame suitable for use
in indicator calculations with the package \code{terradactyl}.The expected format is that used in the Landscape Monitoring Framework (LMF).
}
\examples{
# Using a geodatabase that contains the table PINTERCEPT.
gather_lpi_lmf(dsn = "data_folder/lmf_data.gdb")

# Using a CSV file for PINTERCEPT.
# This argument can also take data frames, if the data have already been read
# in as a data frame.
gather_lpi_lmf(PINTERCEPT = "data_folder/lpi_detail_records.csv")


}
