% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_rangeland_health_lmf}
\alias{gather_rangeland_health_lmf}
\title{Convert LMF-format Interpreting Indicators of RangelandHealth data into a long/tall format.}
\usage{
gather_rangeland_health_lmf(
  dsn = NULL,
  file_type = NULL,
  RANGEHEALTH = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{file_type}{Deprecated. This argument is no longer functional or necessary and is kept for backwards compatibility with legacy code.}

\item{RANGEHEALTH}{Optional data frame or character string. If provided, this must contain the
expected rangeland health intercept data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called RANGEHEALTH (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
Convert LMF-format Interpreting Indicators of RangelandHealth data into a long/tall format.
}
\examples{
# Using a geodatabase that contains the table RANGEHEALTH.
gather_rangeland_health_lmf(dsn = "data_folder/lmf_data.gdb")

# Using a data frame for RANGEHEALTH.
lmf_iirh <- read.csv("data_folder/RANGEHEALTH_detail_records.csv")
gather_rangeland_health_lmf(RANGEHEALTH = lmf_iirh)

}
