% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_rangeland_health_terradat}
\alias{gather_rangeland_health_terradat}
\title{Convert AIM-format Interpreting Indicators of RangelandHealth data into a long/tall format.}
\usage{
gather_rangeland_health_terradat(
  dsn = NULL,
  tblQualHeader = NULL,
  tblQualDetail = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. If provided, this must be the filepath
to a geodatabase which contains the feature classes tblQualHeader and
tblQualDetail, e.g. \code{"C:/DATA/AIM.GDB"}. If this is \code{NULL}, then
the arguments \code{tblQualHeader} and \code{tblQualDetail} are required. Defaults to
\code{NULL}.}

\item{tblQualHeader}{Optional data frame. If provided, this must contain the
expected metadata for the IIRH data. If \code{NULL} then the argument \code{dsn} must be
provided. Defaults to \code{NULL}.}

\item{tblQualDetail}{Optional data frame. If provided, this must contain the
expected IIRH data. If \code{NULL} then the argument
\code{dsn} must be provided. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
Convert AIM-format Interpreting Indicators of RangelandHealth data into a long/tall format.
}
\examples{
# Using a geodatabase that contains the tables tblQualHeader and tblQualDetail
gather_rangeland_health_terradat(dsn = "data_folder/aim_data.gdb")

# Using data frames for tblQualHeader and tblQualDetail
aim_iirhheader <- read.csv("data_folder/iirh_headers.csv")
aim_iirhdetail <- read.csv("data_folder/iirh_detail_records.csv")
gather_gap_terradat(tblQualHeader = aim_iirhheader,
                    tblQualDetail = aim_iirhdetail)



}
