% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_soil_stability}
\alias{gather_soil_stability}
\title{Convert soil stability data into tall, tidy data frame}
\usage{
gather_soil_stability(
  dsn = NULL,
  source,
  file_type = "gdb",
  tblSoilStabDetail = NULL,
  tblSoilStabHeader = NULL,
  SOILDISAG = NULL,
  autoQC = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file
extension) of the geodatabase containing the table of interest. This field
is unnecessary if you supply either both of tblSoilStabDetail and
tblSoilStabHeader (AIM/DIMA/TerrADat) or SOILHORIZON (LMF/NRI).}

\item{source}{Character string. The data source format, can be \code{"AIM"},
\code{"TerrADat"}, \code{"DIMA"}, \code{"LMF"}, or \code{"NRI"} (case independent).}

\item{tblSoilStabDetail}{Dataframe of the data structure tblSoilStabDetail
from the DIMA database with the addition of PrimaryKey and DBKey fields.
Use when data source is AIM, DIMA, or TerrADat; alternately provide \code{dsn}.}

\item{tblSoilStabHeader}{Dataframe of the data structure tblSoilStabHeader
from the DIMA database with the addition of PrimaryKey and DBKey fields.
Use when data source is AIM, DIMA, or TerrADat; alternately provide \code{dsn}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}

\item{SOILHORIZON}{Dataframe of the data structure SOILHORIZON from LMF/NRI
database with the addition of PrimaryKey and DBKey fields. Use when data
source is LMF or NRI; alternately provide \code{dsn}.}

\item{auto_qc}{Logical. Temporarily disabled. If \code{TRUE} then AIM/DIMA/TerrADat data will be
checked for non-unique and orphaned records before doing processing. If any
are found, a warning will be triggered but the gather will still be carried
out. It is strongly recommended that any identified issues be addressed to
avoid incorrect records in the output. Defaults to \code{TRUE}.}
}
\value{
A tall data frame containing soil horizon data.
}
\description{
Given soil stability create a tall format data frame usable by
other terradactyl functions.
}
\examples{
gather_soil_stability(dsn = "Path/To/AIM_Geodatabase.gdb",
                      source = "AIM")
gather_soil_stability(dsn = "Path/To/LMF_Geodatabase.gdb",
                      source = "LMF")

aim_soilstabdetail <- read.csv("Path/To/tblSoilStabDetail.csv")
aim_soilstabheader <- read.csv("Path/To/tblSoilStabHeader.csv")
gather_soil_stability(source = "AIM",
                      tblSoilStabDetail = aim_soilstabdetail,
                      tblSoilStabHeader = aim_soilstabheader)

lmf_horizons <- read.csv("Path/To/SOILHORIZON.csv")
gather_soil_stability(source = "LMF",
                      SOILDISAG = lmf_horizons)
}
