% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_soil_stability}
\alias{gather_soil_stability}
\title{Convert soil stability data into tall, tidy data frame}
\usage{
gather_soil_stability(
  dsn = NULL,
  source,
  file_type = "gdb",
  tblSoilStabDetail = NULL,
  tblSoilStabHeader = NULL,
  SOILDISAG = NULL,
  autoQC = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional character string. If provided, this must be the filepath
to a geodatabase which contains the relevant feature classes. Defaults to \code{NULL}.}

\item{source}{Character string. The data source format, can be \code{"AIM"},
\code{"TerrADat"}, \code{"DIMA"}, \code{"LMF"}, or \code{"NRI"} (case independent).}

\item{file_type}{Deprecated. This argument is no longer functional or necessary and is kept for backwards compatibility with legacy code.}

\item{tblSoilStabDetail}{Optional data frame or character string. If provided, this must contain the
expected soil stability data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called tblSoilStabDetail (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{tblSoilStabHeader}{Optional data frame or character string. If provided, this must contain the
expected metadata for the soil stability data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called tblSoilStabHeader (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{SOILDISAG}{Optional data frame or character string. If provided, this must contain the
expected soil stability intercept data. If \code{NULL} then the argument
\code{dsn} must be provided. If a character string, this must either correspond to the filepath to a CSV or RDATA file containing a table with the data or the name of the feature class in the geodatabase provided as \code{dsn}. If \code{NULL}, the function will attempt to find a feature class called SOILDISAG (making a best guess if there's a partial match) in the \code{dsn} geodatabase. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\value{
A tall data frame containing soil horizon data.
}
\description{
Given soil stability create a tall format data frame usable by
other terradactyl functions. For additional information about
arguments, see the documentation for the functions \code{\link[terradactyl:gather_soil_stability_terradat]{gather_soil_stability_terradat()}} and \code{\link[terradactyl:gather_soil_stability_lmf]{gather_soil_stability_lmf()}}.
}
\details{
The \code{source} argument determines which other arguments are used or ignored.

When \code{source} is one of \code{"AIM"}, \code{"TERRADAT"}, or \code{"DIMA"}
then the arguments \code{tblSoilStabHeader} and \code{tblSoilStabDetail} are both considered.

When \code{source} is \code{"LMF"} then the argument \code{SOILDISAG} is considered.

Regardless of the value of \code{source}, the data sources represented by those other arguments are required. The simplest way to provide them is to provide the filepath to a geodatabase as \code{dsn} with each of those feature classes appearing by the same name as the corresponding argument in that geodatabase.
}
\examples{
gather_soil_stability(dsn = "Path/To/AIM_Geodatabase.gdb",
                      source = "AIM")
gather_soil_stability(dsn = "Path/To/LMF_Geodatabase.gdb",
                      source = "LMF")

aim_soilstabdetail <- read.csv("Path/To/tblSoilStabDetail.csv")
aim_soilstabheader <- read.csv("Path/To/tblSoilStabHeader.csv")
gather_soil_stability(source = "AIM",
                      tblSoilStabDetail = aim_soilstabdetail,
                      tblSoilStabHeader = aim_soilstabheader)

lmf_horizons <- read.csv("Path/To/SOILHORIZON.csv")
gather_soil_stability(source = "LMF",
                      SOILDISAG = lmf_horizons)
}
