% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_soil_stability_lmf}
\alias{gather_soil_stability_lmf}
\title{Convert LMF-format soil stability data into a long/tall format}
\usage{
gather_soil_stability_lmf(
  dsn = NULL,
  file_type = "gdb",
  SOILDISAG = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. If provided, this must be the filepath
to a geodatabase which contains the feature classe SOILDISAG, e.g. \code{"C:/DATA/AIM.GDB"}. If this is \code{NULL}, then
the argument \code{SOILDISAG} is required. Defaults to
\code{NULL}.}

\item{file_type}{Character string. The file extension for the path in \code{dsn}.}

\item{SOILDISAG}{Optional data frame. If provided, this must contain the
expected soil stability data. If \code{NULL} then the argument
\code{dsn} must be provided. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
Convert LMF-format soil stability data into a long/tall format
}
\examples{
# Using a geodatabase that contains the table SOILDISAG.
gather_soil_stability_lmf(dsn = "data_folder/lmf_data.gdb")

# Using a data frame for SOILDISAG.
lmf_soilstability <- read.csv("data_folder/SOILDISAG_detail_records.csv")
gather_soil_stability_lmf(SOILDISAG = lmf_soilstability)

}
