% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{gather_soil_stability_terradat}
\alias{gather_soil_stability_terradat}
\title{Convert AIM-format soil stability data into a long/tall format}
\usage{
gather_soil_stability_terradat(
  dsn = NULL,
  tblSoilStabDetail = NULL,
  tblSoilStabHeader = NULL,
  auto_qc_warnings = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Optional (contingent) character string. If provided, this must be the filepath
to a geodatabase which contains the feature classes tblSoiStabHeader and
tblSoilStabDetail, e.g. \code{"C:/DATA/AIM.GDB"}. If this is \code{NULL}, then
the arguments \code{tblSoilStabHeader} and \code{tblSoilStabDetail} are required. Defaults to
\code{NULL}.}

\item{tblSoilStabDetail}{Optional data frame. If provided, this must contain the
expected metadata for the soil stability data. If \code{NULL} then the argument
\code{dsn} must be provided. Defaults to \code{NULL}.}

\item{tblSoilStabHeader}{Optional data frame. If provided, this must contain the
expected soil stability data. If \code{NULL} then the argument \code{dsn} must be
provided. Defaults to \code{NULL}.}

\item{auto_qc_warnings}{Logical. If \code{TRUE} the function will test the
data and metadata for duplicated records (records which are identical to
each other across all critical variables) and orphaned records (records
which appear in the data but do not have any corresponding metadata
records), either of which can result in erroneous or unexpected outputs.
The number of duplicated and orphaned records will be reported but without
making any changes. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If \code{TRUE} then the function will report back
diagnostic information as console messages while it works. Defaults to
\code{FALSE}.}
}
\description{
Convert AIM-format soil stability data into a long/tall format
}
\examples{
# Using a geodatabase that contains the tables tblGapHeader and tblGapDetail
gather_soil_stability_terradat(dsn = "data_folder/aim_data.gdb")

# Using CSV files for tblSoilStabHeader and tblSoilStabDetail.
# These two arguments can also take data frames, if the data have already
# been read in as data frames.
gather_soil_stability_terradat(tblSoilStabHeader = "data_folder/soilstability_headers.csv",
                               tblSoilStabDetail = "data_folder/soilstability_detail_records.csv")

# Using data frames for tblSoilStabHeader and tblSoilStabDetail
aim_soilstabilityheader <- read.csv("data_folder/soilstability_headers.csv")
aim_soilstabilitydetail <- read.csv("data_folder/soilstability_detail_records.csv")
gather_soil_stability_terradat(tblSoilStabHeader = aim_soilstabilityheader,
                               tblSoilStabDetail = aim_soilstabilitydetail)

}
