% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R, R/species.R
\name{gather_species_inventory}
\alias{gather_species_inventory}
\alias{gather_species_inventory_terradat}
\alias{tall_species}
\alias{gather_species_inventory_lmf}
\alias{species_count}
\title{Convert species inventory data into tall, tidy data frame}
\usage{
gather_species_inventory_terradat(
  dsn = NULL,
  tblSpecRichDetail = NULL,
  tblSpecRichHeader = NULL,
  verbose = FALSE
)

tall_species(species_inventory_detail)

gather_species_inventory_lmf(
  dsn = NULL,
  file_type = "gdb",
  PLANTCENSUS = NULL,
  verbose = FALSE
)

gather_species_inventory(
  dsn = NULL,
  source,
  tblSpecRichDetail = NULL,
  tblSpecRichHeader = NULL,
  PLANTCENSUS = NULL,
  file_type = "gdb",
  autoQC = TRUE,
  verbose = FALSE
)

species_count(
  species_inventory_tall,
  ...,
  indicator_variables = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file
extension) of the geodatabase containing the table of interest. This field
is unnecessary if you supply either both of tblSpecRichDetail and
tblSpecRichHeader (AIM/DIMA/TerrADat) or PLANTCENSUS (LMF/NRI).}

\item{tblSpecRichDetail}{Dataframe of the data structure tblSpecRichDetail
from the DIMA database with the addition of PrimaryKey and DBKey fields.
Use with tblSpecRichHeader when data source is AIM, DIMA, or TerrADat;
alternately provide dsn.}

\item{tblSpecRichHeader}{Dataframe of the data structure tblSpecRichHeader
from the DIMA database with the addition of PrimaryKey and DBKey fields.
Use with tblSpecRichDetail when data source is AIM, DIMA, or TerrADat;
alternately provide dsn.}

\item{PLANTCENSUS}{Dataframe of the data structure PLANTCENSUS from LMF/NRI
database with the addition of PrimaryKey and DBKey fields. Use when data
source is LMF or NRI; alternately provide dsn.}

\item{source}{Character string. The data source format,
\code{"AIM", "TerrADat", "DIMA", "LMF", "NRI"} (case independent).}
}
\value{
A tall data frame containing species inventory data.
}
\description{
Given species inventory data create a tall format data frame
usable by other terradactyl functions.
}
\examples{
gather_species_inventory(dsn = "Path/To/AIM_Geodatabase.gdb",
                         source = "AIM")
gather_species_inventory(dsn = "Path/To/LMF_Geodatabase.gdb",
                         source = "LMF")

aim_specrichdetail <- read.csv("Path/To/tblSpecRichDetail.csv")
aim_specrichheader <- read.csv("Path/To/tblSpecRichHeader.csv")
gather_species_inventory(source = "AIM",
                         tblSpecRichDetail = aim_specrichdetail,
                         tblSpecRichHeader = aim_specrichheader)

lmf_census <- read.csv("Path/To/PLANTCENSUS.csv")
gather_species_inventory(source = "LMF",
                         PLANTCENSUS = lmf_census)
}
\concept{<gather>}
