% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{mean.height}
\alias{mean.height}
\title{Calculate the vegetation height}
\usage{
\method{mean}{height}(lpi.height.tall, method = "mean", omit.zero = FALSE,
  by.line = FALSE, tall = FALSE, ...)
}
\arguments{
\item{method}{Character string. Indicates the type of indicator, \code{"max"}, which yields the average maximum height (of the herbaceous or woody heights) on the plot or \code{"mean"} which yields the mean height by functional group (woody/herbaceous).}

\item{omit.zero}{Logical. If \code{TRUE} the results omit height measurements of \code{0}. Defaults to \code{FALSE}.}

\item{by.line}{Logical. If \code{TRUE} then the results will be calculated on a per-line basis. If \code{FALSE} then the results will be calculated on a per-plot basis. Defaults to \code{FALSE}.}

\item{tall}{Logical. If \code{TRUE} then the returned data frame will be tall rather than wide and will not have observations for non-existent values e.g., if no data fell into a group on a plot, there will be no row for that group on that plot. Defaults to \code{FALSE}.}

\item{...}{Optional bare variable names. One or more variable name from \code{lpi.tall} to calculate percent cover for, e.g. \code{GrowthHabitSub} to calculate percent cover by growth habits or \code{GrowthHabitSub, Duration} to calculate percent cover for categories like perennial forbs, annual graminoids, etc.}

\item{height.tall}{A tall/long-format data frame. Use the data frame \code{"height"} from the \code{gather.lpi()} output.}
}
\description{
Calculate the vegetation height
}
