% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_cover.R
\name{pct_cover}
\alias{pct_cover}
\title{Calculate percent cover from Line-Point Intercept data}
\usage{
pct_cover(
  lpi_tall,
  tall = FALSE,
  hit = "any",
  by_line = FALSE,
  ...,
  indicator_variables = NULL,
  digits = 1,
  verbose = FALSE
)
}
\arguments{
\item{lpi_tall}{Data frame. The long/tall-format line-point intercept data. This must be in the format produced by \code{gather_lpi()}.}

\item{tall}{Logical. If \code{TRUE} then the output will be in a long/tall format. If \code{FALSE} then the output will be wide. Defaults to \code{FALSE}.}

\item{hit}{Character string. This dictates how many and which layers at a pin drop will be used to calculate the percent cover. If \code{"any"} then every record for each pin drop may be considered, which will produce results which sum to more than 100 percent in the case that any pin drop recorded hits in multiple layers. If \code{"first"} then only the first (i.e., farthest from the ground) record at each pin drop will be considered. If \code{"basal"} then only records for which the layer variable contains the value \code{"SoilSurface"} will be considered, which should be one record per pin drop. Defaults to \code{"any"}.}

\item{by_line}{Logical. If \code{TRUE} then the cover will be calculated for each unique combination of values in the PrimaryKey and LineKey variables instead of simply by PrimaryKey. Defaults to \code{FALSE}.}

\item{...}{Optional bare variable names. These must be variable names in \code{lpi_tall} and are used to determine the unique cover types. This is here for legacy support and the use of \code{indicator_variables} is recommended instead.}

\item{indicator_variables}{Optional vector of character strings. These must be variable names in \code{lpi_tall} and are used to determine the unique cover types. For example, using \code{"code"} will result in calculating cover for each unique value in the code variable and using \code{c("Duration", "GrowthHabitSub")} will result in calculating cover for each unique combination of values in the Duration and GrowthHabitSub variables, e.g., PerennialGraminoid and AnnualGraminoid. If \code{NULL} and no variable names have been provided then cover will be calculated per-PrimaryKey and should be 100 percent for each. Defaults to \code{NULL}.}

\item{digits}{Integer. The number of decimal places that the output cover percentages will be rounded to. Values larger than \code{2} are not recommended because they will likely imply false precision. Defaults to \code{1}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
A data frame containing PrimaryKeys and a variable for each cover type containing the percent cover.
}
\description{
Calculate the percent of a transect under cover based on one or more variables in a data frame in the long format produced by \code{gather_lpi()}. Records are grouped by and cover calculated for unique values in the PrimaryKey variable.
}
