% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.nri.R
\name{read_nri_text}
\alias{read_nri_text}
\title{Read NRI txt files using table_name}
\usage{
read_nri_text(table_name, dsn, DBKey = "auto", schema)
}
\arguments{
\item{table_name}{Character string of the nri table names}

\item{dsn}{File path where NRI text files stored}

\item{schema}{Tables and names from NRI Grazing Land Guide}
}
\value{
NRI files with column name assigned
}
\description{
Assign the column header to the NRI text files and save
}
