% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrub.shape.R
\name{sagebrush_shape}
\alias{sagebrush_shape}
\title{Sagebrush Shape}
\usage{
sagebrush_shape(lpi.tall)
}
\arguments{
\item{lpi.tall}{}
}
\description{
Describe the shape characteristics of sagebrush on the plot,
including number of sagebrush hits in each shape type and the predominant shape_
}
