% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_stability.R
\name{soil_stability}
\alias{soil_stability}
\title{Soil Stability Indicator Calculations}
\usage{
soil_stability(soil_stability_tall, all = TRUE, cover = TRUE,
  uncovered = TRUE, all_cover_types = FALSE, tall = FALSE)
}
\arguments{
\item{soil_stability_tall}{Dataframe Gathered soil stability data}

\item{all}{Logical. When \code{TRUE} Calculate soil stability for all samples. Defaults to \code{TRUE}}

\item{cover}{Logical. When \code{TRUE}, calculates soil stability for samples covered byperennial vegetation. Defaults to \code{TRUE}}

\item{all_cover_types}{Logical. When \code{TRUE}, calculates soil stability for each indidual cover type. Defaults to \code{FALSE}}

\item{tall}{Logical. Indicates if output is tall/long or wide. Defaults to \code{TRUE}}

\item{uncovered.}{Logical. When \code{TRUE}, calculates soil stability for samples not covered by perennial vegetation. Defaults to \code{TRUE}}
}
\value{
Dataframe of calculated soil stability values by plot.
}
\description{
Soil Stability Indicator Calculations
}
