% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_inventory.R
\name{gather.species.inventory}
\alias{gather.species.inventory}
\alias{species.count}
\alias{tall.species}
\alias{gather.species.lmf}
\title{Gather Species Inventory}
\usage{
gather.species.inventory(dsn)

species.count(species.inventory.tall, ...)

tall.species(species.inventory.detail)

gather.species.lmf(dsn, file.type = "gdb")
}
\arguments{
\item{dsn}{Character string. The full filepath and filename (including file extension) of the geodatabase containing the table of interest.}
}
\value{
A data frames containing the data from the species inventory data in tall format.
}
\description{
Create a tall species inventory data frame.
}
