% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_join}
\alias{species_join}
\title{Join species characteristics to field data}
\usage{
species_join(
  data,
  data_code = "code",
  species_file,
  species_layer = "tblNationalPlants",
  species_code = "NameCode",
  species_growth_habit_code = "GrowthHabitSub",
  species_duration = "Duration",
  species_property_vars = c("GrowthHabit", "GrowthHabitSub", "Duration", "Family",
    "HigherTaxon", "Nonnative", "Invasive", "Noxious", "SpecialStatus", "Photosynthesis",
    "PJ", "CurrentPLANTSCode"),
  growth_habit_file = "",
  growth_habit_code = "Code",
  overwrite_generic_species = FALSE,
  generic_species_file = "",
  update_species_codes = FALSE,
  by_species_key = FALSE,
  check_species = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame. The source data which contains species codes in the variable specified by \code{species_code}. The only required variable is the one containing the species codes.}

\item{data_code}{Character string. The name of the variable in \code{data} which contains the species codes. Defaults to \code{"code"}.}

\item{species_file}{Data frame or character string. If a character string, it must be the full path (including file extension) to either a CSV containing the species characteristics or a geodatabase containing a table specified with \code{species_layer}. If joining data for Terrestrial AIM indicator calculations using tblNationalPlants and tblStateSpecies, please use \code{species_read_aim()} and provide the output data frame as this argument.}

\item{species_layer}{Character string. If \code{species_file} points to a geodatabase, this must correspond to the name of the feature class or table containing the species characteristics. Defaults to \code{"tblNationalPlants"}.}

\item{species_code}{Character string. The name of the variable in the species characteristics that contains the species codes. Defaults to \code{"NameCode"}.}

\item{species_growth_habit_code}{Character string. The name of the variable in the species characteristics that contains the growth habit assignment values, e.g., \code{"graminoid"}. Defaults to \code{"NameCode"}.}

\item{species_duration}{Character string. The name of the variable in the species characteristics that contains the duration assignment values, e.g., \code{"graminoid"}. Defaults to \code{"NameCode"}.}

\item{species_property_vars}{Vector of character strings. These are the names of variables to carry forward from \code{species_file} and coerce into character strings. If variables in the vector are not found in the species characteristics, the function will skip them silently. Defaults to \code{c("GrowthHabit", "GrowthHabitSub", "Duration", "Family", "HigherTaxon", "Nonnative", "Invasive", "Noxious", "SpecialStatus", "Photosynthesis", "PJ", "CurrentPLANTSCode")}.}

\item{growth_habit_file}{Deprecated. Defaults to \code{""}.}

\item{growth_habit_code}{Deprecated. Defaults to \code{"Code"}.}

\item{overwrite_generic_species}{Logical. If \code{TRUE} then \code{generic_species_file} will be read in and applied. Defaults to \code{FALSE}.}

\item{generic_species_file}{Optional character string. Must specify the full path to a CSV containing generic species information. If this is \code{""} or \code{overwrite_generic_species} is \code{FALSE} then no action will be taken. Defaults to \code{""}.}

\item{update_species_codes}{Logical. If \code{TRUE} then the final output will replace the values in the variable specified with \code{data_code} using values from the variable UpdatedSpeciesCode. This is here for legacy support and is not recommended for ongoing use. Defaults to \code{FALSE}.}

\item{by_species_key}{Logical. If \code{TRUE} then the species information will be joined to the data using the variable SpeciesState. This is here for legacy support and is not compatible with the current structure of Terrestrial AIM Database species information in the. Defaults to \code{FALSE}.}

\item{check_species}{Deprecated. This has no effect on the function because checks for duplicate species are always carried out. Default to \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}
}
\value{
The original data frame provided as \code{data} with variables joined from \code{species_file} according to the species codes.
}
\description{
Joining species characteristics to data can be as simple as a join based on a single variable, but that isn't always the case.
This function handles joining species information to data across a range of possible scenarios and formats.

Some functionality is effectively deprecated due to changes in Terrestrial AIM data formats, but has been maintained for backwards compatibility and handling data collected using earlier schemas.
}
