% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_gdb.R
\name{spp_inventory_calc}
\alias{spp_inventory_calc}
\title{Calculate the standard Terrestrial AIM Database (TerrADat) species indicators}
\usage{
spp_inventory_calc(
  header,
  spp_inventory_tall,
  species_file,
  source,
  species_code_var = "SpeciesCode",
  generic_species_file = NULL,
  verbose = FALSE
)
}
\arguments{
\item{header}{Data frame or character string. The metadata for the plots involved in the calculations, this is used to add the SpeciesState variable by joining with the PrimaryKey variable. If a character string, this must point to a CSV file containing the data.}

\item{spp_inventory_tall}{Data frame or character string. The long/tall-format species inventory data for the plots involved in the calculations. The format must match the output from \code{gather_species_inventory()}. If a character string, this must point to a CSV file containing the data.}

\item{species_file}{Data frame or character string. The species characteristics information. If this is a character string for the filepath to a geodatabase, that geodatabase must contain both the tblNationalPlants and tblStateSpecies tables. Otherwise, this must either be the output from \code{species_read_aim()} or be a character string pointing to a CSV file containing the output from \code{species_read_aim()}.}

\item{source}{Character string. If \code{"terradat"} or \code{"aim"} (case insensitive) then live and "dead" heights will be calculated. Defaults to \code{NULL}.}

\item{species_code_var}{Character string. The name of the variable in the species characteristics that contains the species codes. Defaults to \code{"SpeciesCode"}.}

\item{generic_species_file}{Optional character string. Must specify the full path to a CSV containing generic species information. If this is \code{NULL}. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} the function will produce diagnostic
messages. Defaults to \code{FALSE}.}

\item{digits}{Integer. The number of decimal places that the output values will be rounded to. Values larger than \code{1} are not recommended because they will likely imply false precision. Defaults to \code{1}.}
}
\value{
A data frame matching the format of height indicators in TerrADat.
}
\description{
This function calculates the full set of species-specific indicators that are standard for TerrADat.

This depends on the species characteristics used being those found in tblNationalPlants.
}
