% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terradactyl.R
\docType{package}
\name{terradactyl}
\alias{terradactyl}
\alias{terradactyl-package}
\title{terrdactyl: A package for terrestrial data commons tools}
\description{
This package provides functionality for gathering terrestrial core methods
data, calculating standard variables and indicators, and producing
standard outputs for data repositories and models.
}
\author{
Sarah McCord \email{sarah.mccord@ars.usda.gov}
}
